### 2024-02-16 #6361 @editorialbot
[PRE REVIEW]: SeqMetrics: a unified library for performance metrics calculation in Python
**Submitting author:** @AtrCheema (<a href="http://orcid.org/0000-0002-0031-745X">Ather Abbas</a>)
**Repository:** https://github.com/AtrCheema/SeqMetrics
**Branch with paper.md** (empty if default branch): master
**Version:** v2.0.0
