### 2024-02-16 #6361 @danielskatz
👋 @AtrCheema - thanks for your submission.

Your paper is a bit unusual, however. It doesn't seem to have sections or references. See the [example paper](https://joss.readthedocs.io/en/latest/submitting.html#example-paper-and-bibliography). Please feel free to make changes to your .md file (and perhaps add a .bib file), then use the command `@editorialbot check references` to check the references, and the command `@editorialbot generate pdf` after making changes to the .md file or when the references are right to make a new PDF. editorialbot commands need to be the first entry in a new comment.

As some other minor points, your title is grammatically incorrect: it should be "SeqMetrics: a unified library for performance metrics calculation in Python" (no space before a colon; Python with a P, not a p). Please be sure to change these.

I'm going to mark this as paused until you tell me you finished making changes.
