### 2024-02-17 762145d
added sections references and fig links in paper.md
