+++ paper/paper.md
@@ -1,5 +1,5 @@
 ---
-title: 'SeqMetrics : a unified library for performance metrics calculation in python'
+title: 'SeqMetrics: a unified library for performance metrics calculation in Python'
 tags:
   - Python
   - modeling
@@ -10,6 +10,7 @@ authors:
   - name: Fazila Rubab
     affiliation: 1
   - name: Sara Iftikhar
+    orcid: 0000-0001-7446-6805
     affiliation: 1
   - name: Ather Abbas
     orcid: 0000-0002-0031-745X
@@ -23,11 +24,102 @@ date: 17 Febuary 2024
 bibliography: paper.bib
 ---
 
-Summary
-Current python infrastructure lacks a robust, unified and simplified library for error and performance metrics calculations. The SeqMetrics application responds to this critical need, providing a robust toolkit for evaluating regression and classification models with a comprehensive suite of performance metrics suitable for tabular and time series data. Designed for versatility, the library offers 112 regression and 22 classification metrics through both functional and class-based APIs. The design of the library ensures seamless integration into various coding environments. The web based GUI of SeqMetrics enhances user accessibility, allowing efficient input through data arrays or file imports. It serves to users of varied programming expertise, offering a user-friendly interface for rigorous model assessment. The library prioritizes computational efficiency and has minimal dependency with straightforward pip installation from PyPI. Rigorous testing of Seqmetrics ensures robustness, supported by extensive documentation for effective utilization by people from diverse backgrounds. Overall, SeqMetrics bridges the gap in Python's scientific analysis toolkit, contributing to data analysis.
+# Summary
+Current python infrastructure lacks a robust, unified and simplified library for 
+error and performance metrics calculations. The SeqMetrics application responds 
+to this critical need, providing a robust toolkit for evaluating regression and 
+classification models with a comprehensive suite of performance metrics suitable 
+for tabular and time series data. Designed for versatility, the library offers 
+112 regression and 22 classification metrics through both functional and 
+class-based APIs. The design of the library ensures seamless integration into 
+various coding environments. The web based GUI of SeqMetrics enhances user 
+accessibility, allowing efficient input through data arrays or file imports. 
+It serves to users of varied programming expertise, offering a user-friendly 
+interface for rigorous model assessment. The library prioritizes computational 
+efficiency and has minimal dependency with straightforward pip installation 
+from PyPI. Rigorous testing of Seqmetrics ensures robustness, supported by extensive 
+documentation for effective utilization by people from diverse backgrounds. 
+Overall, SeqMetrics bridges the gap in Python's scientific analysis toolkit, 
+contributing to data analysis.
 
-Problem Statement
-Quantification and analysis of model performance through error and performance metrics is a crucial step in any scientific modeling exercise. However, Python lacks a unified library encompassing performance metrics across diverse fields. Addressing this gap is imperative to provide researchers a unified platform for comprehensive model evaluation, streamlining their computational workflows and enhancing accuracy across various domains. The SeqMetrics application addresses the critical need for a robust and versatile toolkit for assessing and comparing regression and classification models performance across a spectrum of domains. With a comprehensive suite of performance metrics for sequential (tabular and time series) data, spanning traditional statistical measures to specialized atmospheric sciences metrics, the software serves as a valuable resource for researchers, analysts, and practitioners in fields such as hydrology, finance, and engineering.   By providing a unified platform for evaluating model performance through a diverse set of metrics, the application facilitates the rigorous validation and optimization of regression and classification models, contributing to informed decision-making processes and ensuring the reliability of predictive modeling in complex and dynamic systems.
[…]
