### 2024-02-17 2d9808d
updated .bib and .md files
