### 2024-02-17 df09841
updated hyperlines for figs in paper.md
