### 2024-02-17 #6361 @AtrCheema
@danielskatz Thanks for your feedback. I am finished with making changes that you have recommended. The sections are visible now. Can you allow the further processing now?
