### 2024-02-17 #6361 @danielskatz
👋 @AtrCheema - you still need to fix the DOI in your one reference - see the PDF, where there's clearly an extra prefix, which this command should show. Also, a paper with only one reference seems odd to me, but we can see what reviewers suggest. I'll start the process.
