### 2024-02-26 7fe92b6
added variability ratio
