+++ SeqMetrics/_rgr.py
@@ -181,7 +181,7 @@ class RegressionMetrics(Metrics):
         return kge_bound(true=self.true, predicted=self.predicted, preprocess=False)
 
     def kge_mod(self):
-        return kge(true=self.true, predicted=self.predicted, preprocess=False)
+        return kge_mod(true=self.true, predicted=self.predicted, preprocess=False)
 
     def kge_np(self):
         return kge_np(true=self.true, predicted=self.predicted, preprocess=False)
@@ -423,6 +423,9 @@ class RegressionMetrics(Metrics):
 
     def wmape(self) -> float:
         return wmape(true=self.true, predicted=self.predicted, preprocess=False)
+    
+    def variability_ratio(self) -> float:
+        return variability_ratio(true=self.true, predicted=self.predicted, preprocess=False)
 
 
 def post_process_kge(cc, alpha, beta, return_all=False):
@@ -515,7 +518,7 @@ def nse(true, predicted, preprocess: bool = True) -> float:
 def nse_alpha(true, predicted, preprocess: bool = True) -> float:
     """
     Alpha decomposition of the NSE, see `Gupta_ et al. 2009 <https://doi.org/10.1029/97WR03495>`_
-    used in kratzert et al., 2018
+    used in `kratzert et al., 2018 <>`_
 
     Returns
     -------
@@ -746,15 +749,18 @@ def adjusted_r2(true, predicted, preprocess: bool = True) -> float:
 
 def kge(true, predicted, preprocess: bool = True, return_all=False):
     """
-    Kling-Gupta Efficiency
-    Gupta, Kling, Yilmaz, Martinez, 2009, Decomposition of the mean squared error and NSE performance
-     criteria: Implications for improving hydrological modelling
+    Kling-Gupta Efficiency following `Gupta_ et al. 2009 <https://doi.org/10.1016/j.jhydrol.2009.08.003>`_.
+
 
     output:
+        If return_all is True, it returns a numpy array of shape (4, ) containing
+        kge, cc, alpha, beta. Otherwise, it returns kge.
+
         kge: Kling-Gupta Efficiency
         cc: correlation
         alpha: ratio of the standard deviation
         beta: ratio of the mean
+
     Parameters
     ----------
     true :
@@ -783,10 +789,8 @@ def kge(true, predicted, preprocess: bool = True, return_all=False):
 
 def kge_bound(true, predicted, preprocess: bool = True) -> float:
     """
-    Bounded Version of the Original Kling-Gupta Efficiency_
-
-    .. _Efficiency:
-        https://iahs.info/uploads/dms/13614.21--211-219-41-MATHEVET.pdf
+    Bounded Version of the Original Kling-Gupta Efficiency after
+    `Mathevet et al. 2006 <https://iahs.info/uploads/dms/13614.21--211-219-41-MATHEVET.pdf>`_.
 
     Parameters
     ----------
@@ -815,10 +819,7 @@ def kge_bound(true, predicted, preprocess: bool = True) -> float:
 
 def kge_mod(true, predicted, preprocess: bool = True, return_all=False):
     """
-    Modified Kling-Gupta Efficiency_ .
-
-    .. _Efficiency:
-        https://doi.org/10.1016/j.jhydrol.2012.01.011
+    Modified Kling-Gupta Efficiency after `Kling et al. 2012 <https://doi.org/10.1016/j.jhydrol.2012.01.011>`_.
 
     Parameters
     ----------
@@ -858,19 +859,7 @@ def kge_mod(true, predicted, preprocess: bool = True, return_all=False):
 
 def kge_np(true, predicted, preprocess: bool = True, return_all=False):
     """
-    Non parametric Kling-Gupta Efficiency
-
-    output:
-        kge: Kling-Gupta Efficiency
-        cc: correlation
[…]
