### 2024-02-29 #6361 @mstimberg
Hi again, @AtrCheema. I've contacted a few potential reviewers over email and waiting for their response. From a cursory look at your paper/project, I think that reviewers will ask about comparisons with existing packages  – i.e., what does this package offer that e.g. [scikit-learn's metrics package](https://scikit-learn.org/stable/modules/classes.html#module-sklearn.metrics), [Kera's metrics package](https://www.tensorflow.org/api_docs/python/tf/keras/metrics), or the [torchmetrics package](https://lightning.ai/docs/torchmetrics/stable/) do not?
