### 2024-03-01 142ac37
remove repetitious call to preprocess function
