+++ SeqMetrics/_rgr.py
@@ -175,7 +175,7 @@ class RegressionMetrics(Metrics):
                             preprocess=False)
 
     def kge(self):
-        return kge(true=self.true, predicted=self.predicted)
+        return kge(true=self.true, predicted=self.predicted, preprocess=False)
 
     def kge_bound(self) -> float:
         return kge_bound(true=self.true, predicted=self.predicted, preprocess=False)
@@ -326,10 +326,10 @@ class RegressionMetrics(Metrics):
         return nse_bound(true=self.true, predicted=self.predicted, preprocess=False)
 
     def log_nse(self, epsilon=0.0) -> float:
-        return log_nse(true=self.true, predicted=self.predicted, epsilon=epsilon)
+        return log_nse(true=self.true, predicted=self.predicted, epsilon=epsilon, preprocess=False)
 
     def log_prob(self) -> float:
-        return log_prob(true=self.true, predicted=self.predicted)
+        return log_prob(true=self.true, predicted=self.predicted, preprocess=False)
 
     def pbias(self) -> float:
         return pbias(true=self.true, predicted=self.predicted, preprocess=False)
@@ -404,7 +404,7 @@ class RegressionMetrics(Metrics):
         return sse(true=self.true, predicted=self.predicted, preprocess=False)
 
     def std_ratio(self, **kwargs) -> float:
-        return std_ratio(true=self.true, predicted=self.predicted, **kwargs)
+        return std_ratio(true=self.true, predicted=self.predicted, preprocess=False, **kwargs)
 
     def umbrae(self, benchmark: np.ndarray = None):
         return umbrae(true=self.true, predicted=self.predicted, preprocess=False, benchmark=benchmark)
@@ -423,7 +423,7 @@ class RegressionMetrics(Metrics):
 
     def wmape(self) -> float:
         return wmape(true=self.true, predicted=self.predicted, preprocess=False)
-    
+
     def variability_ratio(self) -> float:
         return variability_ratio(true=self.true, predicted=self.predicted, preprocess=False)
 
@@ -436,7 +436,8 @@ def post_process_kge(cc, alpha, beta, return_all=False):
         return kge_
 
 
-def r2(true, predicted, preprocess: bool = True) -> float:
+def r2(true, predicted, preprocess: bool = True,
+       **process_kws) -> float:
     """
     R2 is a statistical measure of how well the regression line approximates the actual data.
     Quantifies the percent of variation in the response that the 'model'
@@ -465,7 +466,7 @@ def r2(true, predicted, preprocess: bool = True) -> float:
     >>> p = np.random.random(10)
     >>> r2(t, p)
     """
-    true, predicted = maybe_preprocess(preprocess, true, predicted, METRIC_TYPES['r2'])
+    true, predicted = maybe_preprocess(preprocess, true, predicted, METRIC_TYPES['r2'], **process_kws)
 
     zx = (true - np.mean(true)) / np.std(true, ddof=1)
     zy = (predicted - np.mean(predicted)) / np.std(predicted, ddof=1)
@@ -473,7 +474,8 @@ def r2(true, predicted, preprocess: bool = True) -> float:
     return float(r ** 2)
 
 
-def nse(true, predicted, preprocess: bool = True) -> float:
+def nse(true, predicted, preprocess: bool = True,
+        **process_kws) -> float:
     """Nash-Sutcliff Efficiency.
 
     The Nash-Sutcliffe efficiency (NSE) is a normalized statistic that determines
[…]
