### 2024-03-01 7c94844
make use of treat_arrays function
