+++ tests/test_rgr.py
@@ -1,6 +1,6 @@
 import os
 import unittest
-import site   # so that ai4water directory is in path
+import site  # so that ai4water directory is in path
 
 ai4_dir = os.path.dirname(os.path.dirname(os.path.abspath(__file__)))
 site.addsitedir(ai4_dir)
@@ -116,12 +116,11 @@ from SeqMetrics import watt_m
 from SeqMetrics import wmape
 from SeqMetrics import variability_ratio
 
-
 t = np.random.random((20, 1))
 p = np.random.random((20, 1))
 
-ts = pd.Series(np.random.random((20, )))
-ps = pd.Series(np.random.random((20, )))
+ts = pd.Series(np.random.random((20,)))
+ps = pd.Series(np.random.random((20,)))
 tdf = pd.DataFrame(np.random.random((20, 1)))
 pdf = pd.DataFrame(np.random.random((20, 1)))
 
@@ -131,7 +130,7 @@ erdf = RegressionMetrics(tdf, pdf)
 
 all_errors = er.calculate_all()
 
-not_metrics = ['calculate_all', 
+not_metrics = ['calculate_all',
                "treat_arrays",
                "scale_free_metrics",
                "scale_dependent_metrics",
@@ -139,11 +138,11 @@ not_metrics = ['calculate_all',
                "relative_metrics",
                "percentage_metrics"]
 
-
 random_state = np.random.RandomState(seed=313)
 
 t11 = random_state.random(100)
 p11 = random_state.random(100)
+
 
 class test_errors(unittest.TestCase):
 
@@ -184,25 +183,25 @@ class test_errors(unittest.TestCase):
     def test_mare(self):
         # https://support.numxl.com/hc/en-us/articles/115001223363-MRAE-Mean-Relative-Absolute-Error
         data = np.array(
-            [[-2.9, 	-2.95],
-             [-2.83, 	-2.7],
-             [-0.95, 	-1.00],
-             [-0.88, 	-0.68],
-             [1.21,	1.50],
-             [-1.67, 	-1.00],
-             [0.83, 	0.90],
-             [-0.27, 	-0.37],
-             [1.36, 	1.26],
-             [-0.34, 	-0.54],
-             [0.48, 	0.58],
-             [-2.83, 	-2.13],
-             [-0.95, 	-0.75],
-             [-0.88, 	-0.89],
-             [1.21, 	1.25],
-             [-1.67, 	-1.65],
-             [-2.99, 	-3.20],
-             [1.24, 	1.29],
-             [0.64, 	0.60]]
+            [[-2.9, -2.95],
+             [-2.83, -2.7],
+             [-0.95, -1.00],
+             [-0.88, -0.68],
+             [1.21, 1.50],
+             [-1.67, -1.00],
+             [0.83, 0.90],
+             [-0.27, -0.37],
+             [1.36, 1.26],
+             [-0.34, -0.54],
+             [0.48, 0.58],
+             [-2.83, -2.13],
+             [-0.95, -0.75],
+             [-0.88, -0.89],
+             [1.21, 1.25],
+             [-1.67, -1.65],
+             [-2.99, -3.20],
+             [1.24, 1.29],
+             [0.64, 0.60]]
         )
         errs = RegressionMetrics(data[:, 0], data[:, 1])
         np.testing.assert_almost_equal(0.348, errs.mrae(), 2)
@@ -233,303 +232,378 @@ class test_errors(unittest.TestCase):
         new_r2 = r2(t11, p11)
         assert np.allclose(new_r2, 0.0003276772244559177)
         return
+
     def test_nse(self):
         new_nse = nse(t11, p11)
         assert np.allclose(new_nse, -1.068372251749874)
         return
+
[…]
