+++ SeqMetrics/_main.py
@@ -5,20 +5,20 @@ import warnings
 import numpy as np
 from typing import Union
 
-from .utils import maybe_preprocess
 from .utils import features
+from .utils import maybe_preprocess
+
 
 # TODO remove repeated calculation of mse, std, mean etc
 # TODO make weights, class attribute
 # TODO standardized residual sum of squares
-# TODO treat_metrics is not being used!
 # http://documentation.sas.com/?cdcId=fscdc&cdcVersion=15.1&docsetId=fsug&docsetTarget=n1sm8nk3229ttun187529xtkbtpu.htm&locale=en
 # https://arxiv.org/ftp/arxiv/papers/1809/1809.03006.pdf
 # https://www.researchgate.net/profile/Mark-Tschopp/publication/322147437_Quantifying_Similarity_and_Distance_Measures_for_Vector-Based_Datasets_Histograms_Signals_and_Probability_Distribution_Functions/links/5a48089ca6fdcce1971c8142/Quantifying-Similarity-and-Distance-Measures-for-Vector-Based-Datasets-Histograms-Signals-and-Probability-Distribution-Functions.pdf
 # Jeffreys Divergence
 # kullback-Leibler divergence
 # Peak flow ratio https://hess.copernicus.org/articles/24/869/2020/
-# Legates׳s coefficient of efficiency
+# Legates?s coefficient of efficiency
 # outliear percentage : pysteps
 # mean squared error skill score, mean absolute error skill score, https://doi.org/10.1016/j.ijforecast.2018.11.010
 # root mean quartic error, Kolmogorov–Smirnov test integral, OVERPer, Rényi entropy,
@@ -67,6 +67,8 @@ class Metrics(object):
             replace_inf: Union[int, float, None] = None,
             remove_zero: bool = False,
             remove_neg: bool = False,
+            remove_nan: bool = True,
+            remove_inf: bool = True,
             metric_type: str = 'regression',
             np_errstate: dict = None,
     ):
@@ -97,11 +99,23 @@ class Metrics(object):
         global ERR_STATE
 
         self.metric_type = metric_type
-        self.true, self.predicted = maybe_preprocess(True, true, predicted, metric_type)
+        self.true, self.predicted = maybe_preprocess(
+            preprocess=True,
+            true=true,
+            predicted=predicted,
+            metric_type=metric_type,
+            remove_nan=remove_nan,
+            replace_nan=replace_nan,
+            remove_zero=remove_zero,
+            remove_neg=remove_neg,
+            replace_inf=replace_inf,
+            remove_inf=remove_inf,
+            )
         self.replace_nan = replace_nan
         self.replace_inf = replace_inf
         self.remove_zero = remove_zero
         self.remove_neg = remove_neg
+        self.remove_nan = remove_nan
         if np_errstate is None:
             np_errstate = {}
         self.err_state = np_errstate
@@ -393,113 +407,6 @@ class Metrics(object):
     def composite_metrics(self):
         pass
 
-    def treat_values(self):
-        """
-        This function is applied by default at the start/at the time of initiating
-        the class. However, it can be used any time after that. This can be handy
-        if we want to calculate error first by ignoring nan and then by no ignoring
-        nan. Adopting from HydroErr_ . Removes the nan, negative, and inf values 
-        in two numpy arrays
-
-        .. _HydroErr:
-            https://github.com/BYU-Hydroinformatics/HydroErr/blob/master/HydroErr/HydroErr.py#L6210
-        """
-        sim_copy = np.copy(self.predicted)
-        obs_copy = np.copy(self.true)
-
-        # Treat missing data in observed_array and simulated_array, rows in simulated_array or
[…]

[…]
