### 2024-03-01 1ee8cab
added comparison with other libraries in paper
