### 2024-03-02 8156bae
change preprocess to treat_arrays and improved docs
