+++ SeqMetrics/_rgr.py
@@ -5,8 +5,7 @@ from typing import Union
 import numpy as np
 from scipy.stats import gmean, kendalltau
 
-from .utils import maybe_preprocess
-from .utils import METRIC_TYPES
+from .utils import maybe_treat_arrays
 from .utils import _geometric_mean, _mean_tweedie_deviance, _foo, list_subclass_methods
 from ._main import Metrics, EPS, ERR_STATE
 
@@ -77,355 +76,355 @@ class RegressionMetrics(Metrics):
                 'mase', 'r2_score']
 
     def abs_pbias(self) -> float:
-        return abs_pbias(true=self.true, predicted=self.predicted, preprocess=False)
+        return abs_pbias(true=self.true, predicted=self.predicted, treat_arrays=False)
 
     def acc(self) -> float:
-        return acc(true=self.true, predicted=self.predicted, preprocess=False)
+        return acc(true=self.true, predicted=self.predicted, treat_arrays=False)
 
     def adjusted_r2(self) -> float:
-        return adjusted_r2(true=self.true, predicted=self.predicted, preprocess=False)
+        return adjusted_r2(true=self.true, predicted=self.predicted, treat_arrays=False)
 
     def agreement_index(self) -> float:
-        return agreement_index(true=self.true, predicted=self.predicted, preprocess=False)
+        return agreement_index(true=self.true, predicted=self.predicted, treat_arrays=False)
 
     def aic(self, p=1) -> float:
-        return aic(true=self.true, predicted=self.predicted, preprocess=False, p=p)
+        return aic(true=self.true, predicted=self.predicted, treat_arrays=False, p=p)
 
     def aitchison(self, center='mean') -> float:
-        return aitchison(true=self.true, predicted=self.predicted, preprocess=False,
+        return aitchison(true=self.true, predicted=self.predicted, treat_arrays=False,
                          center=center)
 
     def amemiya_adj_r2(self) -> float:
-        return amemiya_adj_r2(true=self.true, predicted=self.predicted, preprocess=False)
+        return amemiya_adj_r2(true=self.true, predicted=self.predicted, treat_arrays=False)
 
     def amemiya_pred_criterion(self) -> float:
         return amemiya_pred_criterion(true=self.true, predicted=self.predicted,
-                                      preprocess=False)
+                                      treat_arrays=False)
 
     def bias(self) -> float:
-        return bias(true=self.true, predicted=self.predicted, preprocess=False)
+        return bias(true=self.true, predicted=self.predicted, treat_arrays=False)
 
     def bic(self, p=1) -> float:
-        return bic(true=self.true, predicted=self.predicted, preprocess=False, p=p)
+        return bic(true=self.true, predicted=self.predicted, treat_arrays=False, p=p)
 
     def brier_score(self) -> float:
-        return brier_score(true=self.true, predicted=self.predicted, preprocess=False)
+        return brier_score(true=self.true, predicted=self.predicted, treat_arrays=False)
 
     def corr_coeff(self) -> float:
-        return corr_coeff(true=self.true, predicted=self.predicted, preprocess=False)
+        return corr_coeff(true=self.true, predicted=self.predicted, treat_arrays=False)
 
     def covariance(self) -> float:
-        return covariance(true=self.true, predicted=self.predicted, preprocess=False)
+        return covariance(true=self.true, predicted=self.predicted, treat_arrays=False)
 
     def cronbach_alpha(self) -> float:
-        return cronbach_alpha(true=self.true, predicted=self.predicted, preprocess=False)
+        return cronbach_alpha(true=self.true, predicted=self.predicted, treat_arrays=False)
 
     def centered_rms_dev(self) -> float:
-        return centered_rms_dev(true=self.true, predicted=self.predicted, preprocess=False)
[…]
