### 2024-03-02 f1b8858
added compatibility with xr DataArray torch and tf tensors
