### 2024-03-02 #6361 @AtrCheema
Hi, @mstimberg Thank you for your feedback and for highlighting the lack of comparison in the submitted paper. One significant distinction of SeqMetrics from other libraries lies in its extensive coverage of performance metrics for 1-dimensional numerical data. Existing Python libraries offer only a limited number of metrics in comparison. For instance, the metrics sub-module from [Keras](https://www.tensorflow.org/api_docs/python/tf/keras/metrics) contains only 24 metrics, while [scikit-learn's](https://scikit-learn.org/stable/modules/classes.html#module-sklearn.metrics) metrics module covers 45 metrics. [Torchmetrics ](https://lightning.ai/docs/torchmetrics/stable/all-metrics.html) library, although contains 100+ metrics, however, it provides only 48 such metrics which are intended for 1-dimensional data. In contrast, SeqMetrics covers a comprehensive list of 126 metrics spanning diverse fields.

Another noteworthy feature of SeqMetrics is the availability of a user-friendly and intuitive web-based GUI hosted on [Streamlit](https://seqmetrics.streamlit.app/). This facilitates users with no programming background in seamlessly calculating any of these 126 metrics. We have accordingly modified the paper.md file to reflect these key distinctions.
