### 2024-03-07 #6450 @editorialbot
[REVIEW]: SeqMetrics : a unified library for performance metrics calculation in python
**Submitting author:** @AtrCheema (<a href="http://orcid.org/0000-0002-0031-745X">Ather Abbas</a>)
**Repository:** https://github.com/AtrCheema/SeqMetrics
**Branch with paper.md** (empty if default branch): master
**Version:** v2.0.0
