### 2024-03-07 #6450 @mstimberg
👋🏼  @AtrCheema, @FATelarico, @y1my1, @SkafteNicki  this is the review thread for the paper. All of our communications will happen here from now on.

There are additional guidelines in the [message at the start of this issue](https://github.com/openjournals/joss-reviews/issues/6450#issue-2173755524).
