### 2024-03-07 #6450 @SkafteNicki
Going to preface my review by saying that I am the maintainer of [Torchmetrics](https://github.com/Lightning-AI/torchmetrics), which is being reference in this paper. The TM team welcomes more libraries in within the field of evaluating machine learning models, as we consider this paramount for the field of machine learning to move forward. Also we do not see SegMetrics to be a library in direct competition as the difference in computational backend (pytorch for TM vs numpy for SegMetrics) makes each package suitable for different researchers.

Overall I am satisfied with the paper as it is now. SegMetrics is a nice software package that can be used to calculate a large range of metric on 1D data. It is therefore narrow in scope, but that also means that it can be great at what it does (it definitely seems faster than torchmetrics for calculating a lot of metric in one go). It has a simple and consistent interface and is easy to use. The paper have a clear problem statement, relevant references and a explanation of the API.

However, my main concern is regarding the robustness of the package which is a large claim from the authors throughout the paper. I have laid out my full review in this issue: https://github.com/AtrCheema/SeqMetrics/issues/3, with proposed changes. There are a few breaking points for me at the moment to recommend this paper being accepted to JOSS.
