### 2024-03-08 #6450 @FATelarico
I concur with @SkafteNicki that the submission is almost good to go. However, I have to subscribe to some of the concerns he raised in his review (and the associated issue). Moreover, I have the following, short comments, to make.

## Functionality

### Installation
As I have encountered significant issue with this sort of flaws in the past, I invite the authors to check the entire software on the newest stable release of Python 3 and update all the main packages (especially `numpy`, of which most people probably have a much more recent version).

- [ ] Outdated dependendencies

Personally, I had no problems running the programm on Python (downgraded to 3.7) under both Ubuntu 22. But it would not run on Windows 10 (64 bit) with Python 3.12.2 without downgrading. However, compatibility should be verified after the program is fully updated and the relevant information should be appended to the `README.md` file only then.

- [ ] OS comptibility unclear

### Functionality

I am not completely sure this is the right heading under which to put this comment, but it relates to 'claims' the paper makes.  In fact, I did not see satisfactory indications of the tests' robustness. @SkafteNicki wrote extensively and better than I could about it

- [ ] Robustness claims (see also [this issue](https://github.com/AtrCheema/SeqMetrics/issues/3)

## Documentation

### Community guidelines
There are ready-made templates to add community guidelines. Consider just copy-pasting them and adapting the content to your desires. For instance: [https://bttger.github.io/contributing-gen-web/](https://bttger.github.io/contributing-gen-web/) which is based on [`contributing-gen`](https://github.com/bttger/contributing-gen)

Consider adding an _Installation for contributors_ heading for quick reference if in agreement with your intended policy.

- [ ] Lack of contributing guidelines

## Software paper

### State of the field
Content referrable under this point is contained in rows 24-31 of the Statement of need section. I would like the authors to consider shortening these passages and add a separate heading explictly dedicated to comparing their sofware to `Keras`, `scikit-learn`, `Torchmetrics`, `forecasting_metrics`, `hydroeval`, and other. They do not need to necessarily consider _all_ of them, but at least the most widely used.

In particular, the paper would benefit from a clear description of (some of) the use-cases in which `SeqMetrics` is _technically_ preferable to existing alternatives as opposed to application in which its main added value is the GUI. For instance, the emphasis here is clearly on tabular and time-series, one-dimension data. But reading the paper, at times, one may forget it.

- [ ] Shallow comparison

### Quality of writing
The language and style satisfy the standards of academic writing. However, I suspend the checking of this box until the paper is complete.

### References
Five references seem too few for a paper that should help `SeqMetrics` stand out in a rather crowded field. Even if abovementioned suggestions to include additional sections are rejected, this issue ought to be settled through a more intenst dialogue with existing tools.

- [ ] Partial engagement with the field

## _Postface to any review_
The present comments are to intended as _invitations_ to realise some edits and motivated rejections can lead to constructive arguments in some cases.
