### 2024-03-21 a9925db
Adding documentation of class based API
