+++ SeqMetrics/_rgr.py
@@ -76,356 +76,2057 @@ class RegressionMetrics(Metrics):
                 'mase', 'r2_score']
 
     def abs_pbias(self) -> float:
+        """Absolute Percent bias
+
+        Examples
+        ---------
+        >>> import numpy as np
+        >>> from SeqMetrics import RegressionMetrics
+        >>> t = np.random.random(10)
+        >>> p = np.random.random(10)
+        >>> metrics= RegressionMetrics(t, p)
+        >>> metrics.abs_pbias()
+        """
         return abs_pbias(true=self.true, predicted=self.predicted, treat_arrays=False)
 
     def acc(self) -> float:
+        """Anomaly correction coefficient. See Langland_ et al., 2012; Miyakoda_ et al., 1972
+        and Murphy_ et al., 1989.
+
+        .. _Langland:
+            https://doi.org/10.3402/tellusa.v64i0.17531
+
+        .. _Miyakoda:
+
+        .. _Murphy:
+            https://doi.org/10.1080/02723646.1972.10642213
+
+        Examples
+        ---------
+        >>> import numpy as np
+        >>> from SeqMetrics import RegressionMetrics
+        >>> t = np.random.random(10)
+        >>> p = np.random.random(10)
+        >>> metrics= RegressionMetrics(t, p)
+        >>> metrics.acc()
+        """
         return acc(true=self.true, predicted=self.predicted, treat_arrays=False)
 
     def adjusted_r2(self) -> float:
+        """Adjusted R squared.
+
+        Examples
+        ---------
+        >>> import numpy as np
+        >>> from SeqMetrics import RegressionMetrics
+        >>> t = np.random.random(10)
+        >>> p = np.random.random(10)
+        >>> metrics= RegressionMetrics(t, p)
+        >>> metrics.adjusted_r2()
+        """
         return adjusted_r2(true=self.true, predicted=self.predicted, treat_arrays=False)
 
     def agreement_index(self) -> float:
+        """
+        Agreement Index (d) developed by Willmott_, 1981.
+
+        It detects additive and pro-portional differences in the observed and
+        simulated means and vari-ances Moriasi_ et al., 2015. It is overly sensitive
+        to extreme values due to the squared differences_. It can also be used
+        as a substitute for R2 to identify the degree to which model predic-tions
+        are error-free.
+
+        .. math::
+            d = 1 - \\frac{\\sum_{i=1}^{N}(e_{i} - s_{i})^2}{\\sum_{i=1}^{N}(\\left | s_{i} - \\bar{e}
+             \\right | + \\left | e_{i} - \\bar{e} \\right |)^2}
+
+        .. _differences:
+            Legates and McCabe, 199
+
+        .. _Willmott:
+            https://doi.org/10.1080/02723646.1981.10642213
+
+        .. _Moriasi:
+            https://doi.org/10.13031/trans.58.10715
+
+        Examples
+        ---------
+        >>> import numpy as np
+        >>> from SeqMetrics import RegressionMetrics
+        >>> t = np.random.random(10)
+        >>> p = np.random.random(10)
+        >>> metrics= RegressionMetrics(t, p)
+        >>> metrics.agreement_index()
+        """
         return agreement_index(true=self.true, predicted=self.predicted, treat_arrays=False)
 
     def aic(self, p=1) -> float:
+        """
+        Akaike_ Information Criterion. Modifying from this source_
+
+        .. _Akaike:
+            https://doi.org/10.1007/978-1-4612-1694-0_15
+
+        .. _source:
+            https://github.com/UBC-MDS/RegscorePy/blob/master/RegscorePy/aic.py
+
+        Examples
+        ---------
+        >>> import numpy as np
+        >>> from SeqMetrics import RegressionMetrics
[…]
