### 2024-03-22 #6450 @y1my1
Overall, this is a good package that may meet some needs of the scientific community. However, I echo most of the comments raised by @SkafteNicki and @FATelarico, especially about the documentation and writing of the paper. @SkafteNicki and @FATelarico have already made great suggestions. These are some just minor issues that may help improve the package.

### Documentation

It's great that the authors provide documentation through [readthedocs](https://seqmetrics.readthedocs.io/en/latest/). The authors put a lot of effort there providing important information about the computation of metrics, like the formulas of some metrics. However, for some of the metrics, the authors just provided a reference there; it would be great if the authors could at least provide a formula that helps users understand the underhood of the computation. And it would be beneficial if the authors could write a concise introduction there.

### Software paper
The writing of the paper follows the academic standard and is most understandable. However, there is room for improvement to make it easier to read. For example, this sentence seems to be not very well-written

> Torchmetrics library, (Detlefsen et al., 2022) although contains 100+ metrics, however, it provides only 48 which are intended for 1-dimensional numerical data.
