### 2024-03-23 d9a9f33
Adding Documentation of class based API (Classification Metrics)
