+++ SeqMetrics/_cls.py
@@ -186,6 +186,18 @@ class ClassificationMetrics(Metrics):
         return None
 
     def cross_entropy(self, epsilon=1e-12) -> float:
+        """
+        Examples
+        --------
+        >>> import numpy as np
+        >>> from SeqMetrics import ClassificationMetrics
+        >>> true = np.array([1, 0, 0, 0])
+        >>> pred = np.array([1, 1, 1, 1])
+        >>> metrics = ClassificationMetrics(true, pred)
+        >>> print(metrics.cross_entropy())
+        Computes cross entropy between targets (encoded as one-hot vectors)
+        and predictions.
+        """
         return cross_entropy(true=self.true, predicted=self.predicted, epsilon=epsilon)
 
     # def hinge_loss(self):
@@ -195,9 +207,41 @@ class ClassificationMetrics(Metrics):
     #     return None
 
     def accuracy(self, normalize: bool = True) -> float:
+        """
+        calculates accuracy
+
+        Examples
+        --------
+        >>> import numpy as np
+        >>> from SeqMetrics import ClassificationMetrics
+        >>> true = np.array([1, 0, 0, 0])
+        >>> pred = np.array([1, 1, 1, 1])
+        >>> metrics = ClassificationMetrics(true, pred)
+        >>> print(metrics.accuracy())
+        """
+
         return accuracy(true=self.true, predicted=self.predicted, normalize=normalize)
 
     def confusion_matrix(self, normalize=False):
+        """
+        calculates confusion matrix
+
+        Examples
+        --------
+        >>> import numpy as np
+        >>> from SeqMetrics import ClassificationMetrics
+        >>> true = np.array([1, 0, 0, 0])
+        >>> pred = np.array([1, 1, 1, 1])
+        >>> metrics = ClassificationMetrics(true, pred)
+        >>> metrics.confusion_matrix()
+
+        multiclass classification
+
+        >>> true = np.random.randint(1, 4, 100)
+        >>> pred = np.random.randint(1, 4, 100)
+        >>> metrics = ClassificationMetrics(true, pred)
+        >>> metrics.confusion_matrix()
+        """
         return confusion_matrix(true=self.true, predicted=self.predicted, normalize=normalize)
 
     def _confusion_matrix(self, normalize=None):
@@ -282,15 +326,82 @@ class ClassificationMetrics(Metrics):
         raise NotImplementedError
 
     def precision(self, average=None):
+        """
+        Returns precision score, also called positive predictive value.
+        It is number of correct positive predictions divided by the total
+        number of positive predictions.
+        TP/(TP+FP)
+
+        Examples
+        --------
+        >>> import numpy as np
+        >>> from SeqMetrics import ClassificationMetrics
+        >>> true = np.array([1, 0, 0, 0])
+        >>> pred = np.array([1, 1, 1, 1])
+        >>> metrics = ClassificationMetrics(true, pred)
+        >>> print(metrics.precision())
+        ...
+        >>> print(metrics.precision(average="macro"))
+        >>> print(metrics.precision(average="weighted"))
+        """
         return precision(true=self.true, predicted=self.predicted, average=average)
 
     def recall(self, average=None):
+        """
+        It is also called sensitivity or true positive rate. It is
+        number of correct positive predictions divided by the total number of positives
+        Formula :
+            True Posivitive / True Positive + False Negative
+
+        Examples
+        --------
+        >>> import numpy as np
+        >>> from SeqMetrics import ClassificationMetrics
+        >>> true = np.array([1, 0, 0, 0])
[…]
