### 2024-04-05 b9accfd
Setting labels for functional and class based API in docs
