### 2024-04-08 4826b8a
Added more regression metrics
