+++ SeqMetrics/_rgr.py
@@ -8,6 +8,7 @@ from scipy.stats import gmean, kendalltau
 from .utils import maybe_treat_arrays
 from .utils import _geometric_mean, _mean_tweedie_deviance, _foo, list_subclass_methods
 from ._main import Metrics, EPS, ERR_STATE
+from torchmetrics.utilities.compute import _safe_xlogy
 
 
 class RegressionMetrics(Metrics):
@@ -2127,6 +2128,98 @@ class RegressionMetrics(Metrics):
         >>> metrics.variability_ratio()
         """
         return variability_ratio(true=self.true, predicted=self.predicted, treat_arrays=False)
+
+    def concordance_corr_coef(self) -> float:
+        """
+        Concordance Correlation Coefficient (CCC)
+
+        Examples
+        ---------
+        >>> import numpy as np
+        >>> from SeqMetrics import RegressionMetrics
+        >>> t = np.random.random(10)
+        >>> p = np.random.random(10)
+        >>> metrics= RegressionMetrics(t, p)
+        >>> metrics.concordance_corr_coef()
+        """
+        return concordance_corr_coef(true= self.true, predicted= self.predicted, treat_arrays=False)
+
+    def critical_success_index(self, threshold= 0.5) -> float:
+        """
+        Critical Success Index (CSI)
+
+        Examples
+        ---------
+        >>> import numpy as np
+        >>> from SeqMetrics import RegressionMetrics
+        >>> t = np.array([0, 1, 1, 0, 0, 1])
+        >>> p = np.array([0, 1, 0, 1, 1, 1])
+        >>> metrics= RegressionMetrics(t, p)
+        >>> metrics.critical_success_index()
+        """
+        return critical_success_index(true= self.true, predicted= self.predicted, treat_arrays=False, threshold= threshold)
+
+
+    def kl_divergence(self) -> float:
+        """
+        Kullback-Leibler Divergence
+
+        Examples
+        ---------
+        >>> import numpy as np
+        >>> from SeqMetrics import RegressionMetrics
+        >>> t = np.array([0.1, 0.2, 0.3, 0.2, 0.2])
+        >>> p = np.array([0.2, 0.2, 0.2, 0.2, 0.2])
+        >>> metrics= RegressionMetrics(t, p)
+        >>> divergence = metrics.kl_divergence()
+        """
+        return kl_divergence(true= self.true, predicted= self.predicted, treat_arrays=False)
+
+    def log_cosh_error(self) -> float:
+        """
+        Log-Cosh Error
+        Examples
+        ---------
+        >>> import numpy as np
+        >>> from SeqMetrics import RegressionMetrics
+        >>> t = np.array([1, 2, 3, 4, 5])
+        >>> p = np.array([1.1, 1.9, 3.1, 4.2, 4.8])
+        >>> metrics= RegressionMetrics(t, p)
+        >>> error = metrics.log_cosh_error()
+        """
+        return log_cosh_error(true= self.true, predicted= self.predicted, treat_arrays=False)
+
+    def minkowski_distance(self, order= 1) -> float:
+        """
+        Minkowski Distance
+
+        Examples
+        ---------
+        >>> import numpy as np
+        >>> from SeqMetrics import RegressionMetrics
+        >>> t = np.array([1, 2, 3, 4, 5])
+        >>> p = np.array([1.1, 1.9, 3.1, 4.2, 4.8])
+        >>> metrics= RegressionMetrics(t, p)
+        >>> distance = metrics.minkowski_distance()
+        """
+        return minkowski_distance(true= self.true, predicted= self.predicted, treat_arrays=False, order= order)
+
[…]
