### 2024-05-05 187332b
added tests for both functional and class based APIS using seqmetrics
