+++ tests/test_rgr.py
@@ -7,10 +7,12 @@ site.addsitedir(ai4_dir)
 
 import numpy as np
 import pandas as pd
-import torch
 
+from sklearn.metrics import max_error, explained_variance_score, mean_absolute_error
+from sklearn.metrics import mean_squared_error, mean_squared_log_error, r2_score
+from sklearn.metrics import median_absolute_error, mean_absolute_percentage_error
 
-from SeqMetrics import r2
+from SeqMetrics import mse as sm_mse
 from SeqMetrics import RegressionMetrics
 from SeqMetrics import nse
 from SeqMetrics import nse_alpha
@@ -18,7 +20,7 @@ from SeqMetrics import nse_beta
 from SeqMetrics import nse_mod
 from SeqMetrics import nse_rel
 from SeqMetrics import nse_bound
-from SeqMetrics import r2_score
+from SeqMetrics import r2_score as sm_r2_score
 from SeqMetrics import adjusted_r2
 from SeqMetrics import kge
 from SeqMetrics import kge_bound
@@ -26,23 +28,22 @@ from SeqMetrics import kge_mod
 from SeqMetrics import kge_np
 from SeqMetrics import log_nse
 from SeqMetrics import corr_coeff
-from SeqMetrics import rmse
-from SeqMetrics import rmsle
-from SeqMetrics import mape
+from SeqMetrics import rmse as sm_rmse
+from SeqMetrics import rmsle as sm_rmsle
+from SeqMetrics import mape as sm_mape
 from SeqMetrics import nrmse
 from SeqMetrics import pbias
 from SeqMetrics import bias
 from SeqMetrics import med_seq_error
-from SeqMetrics import mae
+from SeqMetrics import mae as sm_mae
 from SeqMetrics import abs_pbias
 from SeqMetrics import gmae
 from SeqMetrics import inrse
 from SeqMetrics import irmse
 from SeqMetrics import mase
 from SeqMetrics import mare
-from SeqMetrics import msle
+from SeqMetrics import msle as sm_msle
 from SeqMetrics import covariance
-from SeqMetrics import brier_score
 from SeqMetrics import bic
 from SeqMetrics import sse
 from SeqMetrics import amemiya_pred_criterion
@@ -54,7 +55,7 @@ from SeqMetrics import euclid_distance
 from SeqMetrics import cosine_similarity
 from SeqMetrics import decomposed_mse
 from SeqMetrics import cronbach_alpha
-from SeqMetrics import exp_var_score
+from SeqMetrics import exp_var_score as sm_exp_var_score
 from SeqMetrics import expanded_uncertainty
 from SeqMetrics import fdc_fhv
 from SeqMetrics import fdc_flv
@@ -70,7 +71,7 @@ from SeqMetrics import lm_index
 from SeqMetrics import maape
 from SeqMetrics import mbe
 from SeqMetrics import mbrae
-from SeqMetrics import max_error
+from SeqMetrics import max_error as sm_max_error
 from SeqMetrics import mb_r
 from SeqMetrics import mda
 from SeqMetrics import mde
@@ -81,7 +82,7 @@ from SeqMetrics import mean_bias_error
 from SeqMetrics import mean_var
 from SeqMetrics import mean_poisson_deviance
 from SeqMetrics import mean_gamma_deviance
-from SeqMetrics import median_abs_error
+from SeqMetrics import median_abs_error as sm_median_abs_error
 from SeqMetrics import mle
 from SeqMetrics import mod_agreement_index
 from SeqMetrics import mpe
@@ -150,7 +151,7 @@ random_state = np.random.RandomState(seed=313)
 
 t11 = random_state.random(100)
 p11 = random_state.random(100)
-
[…]
