### 2024-05-05 9908c92
added install options in readme
