### 2024-05-05 260ffab
added instructions to launch streamlit app locally
