+++ tests/test_main.py
@@ -1,5 +1,10 @@
 
+import os
+import site
 import unittest
+
+ai4_dir = os.path.dirname(os.path.dirname(os.path.abspath(__file__)))
+site.addsitedir(ai4_dir)
 
 from SeqMetrics import RegressionMetrics
 
