### 2024-05-05 0b00aea
run on GA for python 3.7 and 3.12
