+++ .github/workflows/py37.yml
@@ -11,7 +11,7 @@ jobs:
     runs-on: ubuntu-latest
     strategy:
       matrix:
-        python-version: [3.7]
+        python-version: [3.7, 3.12]
 
     steps:
     - uses: actions/checkout@v2
