### 2024-05-05 915f593
run GA on windows and ubuntu
