+++ README.md
@@ -29,7 +29,7 @@ or using setup file, go to folder where repo is downloaded
 
 You can also install SeqMetrics with all of its dependencies by making use of ``all`` option
 
-  pip install SeqMetrics[all]
+    pip install SeqMetrics[all]
 
 ## How to Use
 SeqMetrics provides a uniform API for calculation of both regression and classification metrics.
