+++ .github/workflows/py37.yml
@@ -27,7 +27,7 @@ jobs:
         pip install -r dev_requirements.txt
     - name: Test with unittest
       run: |
-        conda install pytest
+        pip install pytest
         pip install coverage
         pip install codecov
         pip install python-coveralls
