### 2024-05-05 8e48b61
run tests on mac os as well
