+++ .github/workflows/pymac.yml
@@ -11,8 +11,8 @@ jobs:
     runs-on: ${{ matrix.os }}
     strategy:
       matrix:
-        os: [ubuntu-latest, windows-latest]
-        python-version: [3.7, 3.12]
+        os: [macos-latest]
+        python-version: [3.8, 3.12]
 
     steps:
     - uses: actions/checkout@v2
