+++ CONTRIBUTING.rst
@@ -1,0 +1,11 @@
+
+Contributions are always welcome and highly appreciated. Please follow the guidelines below.
+
+The contributions are welcome in the form of 
+
+- Bug reports/fixes
+- New performance metrics
+- Documentation improvements
+- New features
+
+The contributors must run the tests and ensure that they pass before submitting a pull request.
