### 2024-05-05 3f43b2f
specify upper numpy version limit and test on GA
