+++ requirements.txt
@@ -1,2 +1,2 @@
-numpy
-scipy
+numpy<=1.26.4
+scipy<=1.13
