+++ .github/workflows/np264.yml
@@ -1,0 +1,38 @@
+name: np264
+on:
+  push:
+    branches: [ master ]
+  pull_request:
+    branches: [ master ]
+
+jobs:
+  build:
+
+    runs-on: ${{ matrix.os }}
+    strategy:
+      matrix:
+        os: [ubuntu-latest]
+        python-version: [3.12]
+
+    steps:
+    - uses: actions/checkout@v2
+    - name: Set up Python ${{ matrix.python-version }}
+      uses: actions/setup-python@v2
+      with:
+        python-version: ${{ matrix.python-version }}
+    - name: Install dependencies
+      run: |
+        python -m pip install --upgrade pip
+        python -m pip install flake8 pytest
+        pip install -r dev_requirements.txt
+        pip install numpy==1.26.4
+    - name: Test with unittest
+      run: |
+        pip install pytest
+        pip install coverage
+        pip install codecov
+        pip install python-coveralls
+
+        coverage run -m unittest tests.test_cls
+        coverage run -m unittest tests.test_rgr
+        coverage run -m unittest tests.test_utils
