### 2024-05-05 89180a3
specify pandas version for numpy lower limit for testing
