+++ .github/workflows/np116.yml
@@ -1,4 +1,4 @@
-name: np116
+name: np117
 on:
   push:
     branches: [ master ]
@@ -25,7 +25,7 @@ jobs:
         python -m pip install --upgrade pip
         python -m pip install flake8 pytest
         pip install -r dev_requirements.txt
-        pip install numpy==1.16.0
+        pip install numpy==1.17.0
         pip install pandas==0.25
     - name: Test with unittest
       run: |
