+++ SeqMetrics/_rgr.py
@@ -3,7 +3,11 @@ from math import sqrt
 from typing import Union
 
 import numpy as np
-from scipy.stats import gmean, kendalltau
+
+try:
+    from scipy.stats import kendalltau
+except (ImportError, ModuleNotFoundError):
+    kendalltau = None, None
 
 from .utils import maybe_treat_arrays
 from .utils import _geometric_mean, _mean_tweedie_deviance, _foo, list_subclass_methods
@@ -55,7 +59,7 @@ class RegressionMetrics(Metrics):
 
         return self._minimal() + [
             'fdc_flv', 'fdc_fhv',
-            'kge', 'kge_np', 'kge_mod', 'kge_bound', 'kgeprime_c2m', 'kgenp_bound',
+            'kge', 'kge_np', 'kge_mod', 'kge_bound', 'kgeprime_bound', 'kgenp_bound',
             'nse', 'nse_alpha', 'nse_beta', 'nse_mod', 'nse_bound']
 
     @staticmethod
@@ -599,7 +603,7 @@ class RegressionMetrics(Metrics):
             - kge_np
             - kge_mod
             - kge_bound
-            - kgeprime_c2m
+            - kgeprime_bound
             - kgenp_bound
             - nse
             - nse_alpha
@@ -690,6 +694,9 @@ class RegressionMetrics(Metrics):
         >>> metrics= RegressionMetrics(t, p)
         >>> metrics.kendaull_tau()
         """
+        if kendalltau is None:
+            raise NotImplementedError("scipy is not installed. Please install scipy to use this method")
+
         return kendaull_tau(true=self.true, predicted=self.predicted, return_p=return_p,
                             treat_arrays=False)
 
@@ -756,7 +763,7 @@ class RegressionMetrics(Metrics):
         """
         return kge_np(true=self.true, predicted=self.predicted, treat_arrays=False)
 
-    def kgeprime_c2m(self) -> float:
+    def kgeprime_bound(self) -> float:
         """
         Bounded Version of the Modified Kling-Gupta Efficiency_
 
@@ -770,10 +777,10 @@ class RegressionMetrics(Metrics):
         >>> t = np.random.random(10)
         >>> p = np.random.random(10)
         >>> metrics= RegressionMetrics(t, p)
-        >>> metrics.kgeprime_c2m()
+        >>> metrics.kgeprime_bound()
 
         """
-        return kgeprime_c2m(true=self.true, predicted=self.predicted, treat_arrays=False)
+        return kgeprime_bound(true=self.true, predicted=self.predicted, treat_arrays=False)
 
     def kgenp_bound(self):
         """
@@ -880,24 +887,6 @@ class RegressionMetrics(Metrics):
         >>> metrics.mape()
         """
         return mape(true=self.true, predicted=self.predicted, treat_arrays=False)
-
-    def mbe(self) -> float:
-        """Mean bias error. This indicator expresses a tendency of model to underestimate (negative value)
-         or overestimate (positive value) global radiation, while the MBE values closest to zero are desirable.
-         The drawback of this test is that it does not show the correct performance when the model presents
-         overestimated and underestimated values at the same time, since overestimation and underestimation
-         values cancel each other_. [1]
-
-         Examples
-         ---------
-         >>> import numpy as np
-         >>> from SeqMetrics import RegressionMetrics
-         >>> t = np.random.random(10)
[…]
