### 2024-05-09 27cf601
added images for using streamlit app in readme and documentation
