### 2024-05-11 c036e8b
increase height of streamlit images
