### 2024-05-11 6ee9008
added tests with multiple inputs
