+++ tests/test_rgr.py
@@ -129,19 +129,7 @@ from SeqMetrics import log_cosh_error as sm_log_cosh_error
 from SeqMetrics import minkowski_distance as sm_minkowski_distance
 from SeqMetrics import tweedie_deviance_score as sm_tweedie_deviance_score
 
-t = np.random.random((20, 1))
-p = np.random.random((20, 1))
-
-ts = pd.Series(np.random.random((20,)))
-ps = pd.Series(np.random.random((20,)))
-tdf = pd.DataFrame(np.random.random((20, 1)))
-pdf = pd.DataFrame(np.random.random((20, 1)))
-
-er = RegressionMetrics(t, p)
-ers = RegressionMetrics(ts, ps)
-erdf = RegressionMetrics(tdf, pdf)
-
-all_errors = er.calculate_all()
+from SeqMetrics.utils import maybe_treat_arrays
 
 not_metrics = ['calculate_all',
                "treat_arrays",
@@ -157,41 +145,50 @@ t11 = random_state.random(100)
 p11 = random_state.random(100)
 metrics = RegressionMetrics(t11, p11)
 
+# check with very large values say > 1e7
+t_large = random_state.random(100) * 1e7
+p_large = random_state.random(100) * 1e7
+metrics_large = RegressionMetrics(t_large, p_large)
+
+# check by inserting NaN values at random places in t11 and p11
+t_nan = t11.copy()
+p_nan = p11.copy()
+nan_indices = random_state.choice(range(100), 10, replace=False)
+t_nan[nan_indices] = np.nan
+nan_indices = random_state.choice(range(100), 10, replace=False)
+p_nan[nan_indices] = np.nan
+metrics_nan = RegressionMetrics(t_nan, p_nan)
+
+# check where some values are negative and some values are positive
+t_neg = random_state.randint(-100, 100, 100)
+p_neg = random_state.randint(-100, 100, 100)
+metrics_neg = RegressionMetrics(t_neg, p_neg)
+
+
 class test_errors(unittest.TestCase):
     
     def test_attrs(self):
         for _attr in not_metrics:
-            assert _attr not in er.all_methods
+            assert _attr not in metrics.all_methods
     
     def test_calculate_all(self):
+        
+        all_errors = metrics.calculate_all()
         assert len(all_errors) > 100
-        for er_name, er_val in all_errors.items():
-            if er_val is not None:
-                er_val = getattr(er, er_name)()
-                self.assertEqual(er_val.__class__.__name__, 'float', f'{er_name} is {er_val}')
-        return
-    
-    def test_calculate_all_series(self):
-        series_errors = ers.calculate_all()
-        assert len(series_errors) > 100
-        for er_name, er_val in series_errors.items():
-            if er_val is not None:
-                er_val = getattr(ers, er_name)()
-                self.assertEqual(er_val.__class__.__name__, 'float', f'{er_name} is {er_val}')
-        return
-    
-    def test_calculate_all_df(self):
-        df_errors = ers.calculate_all()
-        assert len(df_errors) > 100
-        for er_name, er_val in df_errors.items():
-            if er_val is not None:
-                er_val = getattr(erdf, er_name)()
-                self.assertEqual(er_val.__class__.__name__, 'float', f'{er_name} is {er_val}')
-        return
+        assert all([isinstance(val, float) for val in all_errors.values()])
 
-    def test_mrae(self):
[…]
