+++ app.py
@@ -23,7 +23,7 @@ NAME_REG = {'calculate_all' : 'Calculate All Regression Metrics',
             'calculate_hydro_metrics' : 'Calculates All Metrics for Hydrological Data.',
             'adjusted_r2' : 'Adjusted R2',
             'agreement_index': 'Agreement Index',
-            'abs_pbias': 'Absolute Percent Bias',
+            # 'abs_pbias': 'Absolute Percent Bias',
             'amemiya_pred_criterion': 'Amemiya’s Prediction Criterion',
             'amemiya_adj_r2': 'Amemiya’s Adjusted R-squared',
             'aitchison': 'Aitchison Distance',
@@ -59,7 +59,7 @@ NAME_REG = {'calculate_all' : 'Calculate All Regression Metrics',
             'kl_sym': 'Symmetric Kullback-leibler Divergence',
             'lm_index': 'Legate-McCabe Efficiency Index',
             'log_prob': 'Logarithmic Probability Distribution',
-            'log_nse': 'Log Nash-Sutcliffe model Efficiency',
+            # 'log_nse': 'Log Nash-Sutcliffe model Efficiency',
             'maape': 'Mean Arctangent Absolute Percentage Error',
             'mbrae': 'Mean Bounded Relative Absolute Error',
             'max_error': 'Maximum Absolute Error',
