+++ tests/test_rgr.py
@@ -9,7 +9,7 @@ import scipy
 import numpy as np
 import pandas as pd
 
-from scipy.stats import pearsonr, kendalltau
+from scipy.stats import pearsonr, kendalltau, spearmanr
 
 from sklearn.metrics.pairwise import cosine_similarity as sklearn_cos_sim
 from sklearn.metrics import mean_gamma_deviance, mean_poisson_deviance
@@ -31,7 +31,7 @@ from SeqMetrics import kge
 from SeqMetrics import kge_bound
 from SeqMetrics import kge_mod
 from SeqMetrics import kge_np
-from SeqMetrics import log_nse
+# from SeqMetrics import log_nse
 from SeqMetrics import corr_coeff
 from SeqMetrics import rmse as sm_rmse
 from SeqMetrics import rmsle as sm_rmsle
@@ -41,7 +41,7 @@ from SeqMetrics import pbias
 from SeqMetrics import bias
 from SeqMetrics import med_seq_error
 from SeqMetrics import mae as sm_mae
-from SeqMetrics import abs_pbias
+# from SeqMetrics import abs_pbias
 from SeqMetrics import gmae
 from SeqMetrics import inrse
 from SeqMetrics import irmse
@@ -128,6 +128,7 @@ from SeqMetrics import kl_divergence as sm_kl_divergence
 from SeqMetrics import log_cosh_error as sm_log_cosh_error
 from SeqMetrics import minkowski_distance as sm_minkowski_distance
 from SeqMetrics import tweedie_deviance_score as sm_tweedie_deviance_score
+from SeqMetrics import spearmann_corr
 
 from SeqMetrics.utils import maybe_treat_arrays
 
@@ -166,13 +167,12 @@ metrics_neg = RegressionMetrics(t_neg, p_neg)
 
 
 class test_errors(unittest.TestCase):
-    
+
     def test_attrs(self):
         for _attr in not_metrics:
             assert _attr not in metrics.all_methods
-    
+
     def test_calculate_all(self):
-        
         all_errors = metrics.calculate_all()
         assert len(all_errors) > 100
         assert all([isinstance(val, float) for val in all_errors.values()])
@@ -279,57 +279,57 @@ class test_errors(unittest.TestCase):
     def test_r2(self):
         new_r2 = metrics.r2()
         _, _, rvalue, _, _ = scipy.stats.linregress(t11, p11)
-        assert np.allclose(new_r2, rvalue**2)
+        assert np.allclose(new_r2, rvalue ** 2)
 
         new_r2 = metrics_large.r2()
         _, _, rvalue, _, _ = scipy.stats.linregress(t_large, p_large)
-        assert np.allclose(new_r2, rvalue**2)
+        assert np.allclose(new_r2, rvalue ** 2)
 
         new_r2 = metrics_nan.r2()
         t_nan_, p_nan_ = maybe_treat_arrays(True, t_nan, p_nan, 'regression', remove_nan=True)
         _, _, rvalue, _, _ = scipy.stats.linregress(t_nan_, p_nan_)
-        assert np.allclose(new_r2, rvalue**2)
+        assert np.allclose(new_r2, rvalue ** 2)
 
         new_r2 = metrics_neg.r2()
         _, _, rvalue, _, _ = scipy.stats.linregress(t_neg, p_neg)
-        assert np.allclose(new_r2, rvalue**2)
+        assert np.allclose(new_r2, rvalue ** 2)
         return
 
     def test_mse_cls(self):
         new_mse = metrics.mse()
-        sk_mse= mean_squared_error(t11, p11)
+        sk_mse = mean_squared_error(t11, p11)
         assert np.allclose(new_mse, sk_mse)
 
[…]
