+++ SeqMetrics/_rgr.py
@@ -79,19 +79,19 @@ class RegressionMetrics(Metrics):
         return ['r2', 'mape', 'nrmse', 'corr_coeff', 'rmse', 'mae', 'mse', 'mpe',
                 'mase', 'r2_score']
 
-    def abs_pbias(self) -> float:
-        """Absolute Percent bias
-
-        Examples
-        ---------
-        >>> import numpy as np
-        >>> from SeqMetrics import RegressionMetrics
-        >>> t = np.random.random(10)
-        >>> p = np.random.random(10)
-        >>> metrics= RegressionMetrics(t, p)
-        >>> metrics.abs_pbias()
-        """
-        return abs_pbias(true=self.true, predicted=self.predicted, treat_arrays=False)
+    # def abs_pbias(self) -> float:
+    #     """Absolute Percent bias
+    #
+    #     Examples
+    #     ---------
+    #     >>> import numpy as np
+    #     >>> from SeqMetrics import RegressionMetrics
+    #     >>> t = np.random.random(10)
+    #     >>> p = np.random.random(10)
+    #     >>> metrics= RegressionMetrics(t, p)
+    #     >>> metrics.abs_pbias()
+    #     """
+    #     return abs_pbias(true=self.true, predicted=self.predicted, treat_arrays=False)
 
     def acc(self) -> float:
         """Anomaly correction coefficient. See Langland_ et al., 2012; Miyakoda_ et al., 1972
@@ -1499,24 +1499,24 @@ class RegressionMetrics(Metrics):
         """
         return nse_bound(true=self.true, predicted=self.predicted, treat_arrays=False)
 
-    def log_nse(self, epsilon=0.0) -> float:
-        """
-        log Nash-Sutcliffe model efficiency
-
-        .. math::
-            NSE = 1-\\frac{\\sum_{i=1}^{N}(log(e_{i})-log(s_{i}))^2}{\\sum_{i=1}^{N}(log(e_{i})-log(\\bar{e})^2}-1)*-1
-
-            Examples
-        ---------
-        >>> import numpy as np
-        >>> from SeqMetrics import RegressionMetrics
-        >>> t = np.random.random(10)
-        >>> p = np.random.random(10)
-        >>> metrics= RegressionMetrics(t, p)
-        >>> metrics.log_nse()
-
-        """
-        return log_nse(true=self.true, predicted=self.predicted, epsilon=epsilon, treat_arrays=False)
+    # def log_nse(self, epsilon=0.0) -> float:
+    #     """
+    #     log Nash-Sutcliffe model efficiency
+    #
+    #     .. math::
+    #         NSE = 1-\\frac{\\sum_{i=1}^{N}(log(e_{i})-log(s_{i}))^2}{\\sum_{i=1}^{N}(log(e_{i})-log(\\bar{e})^2}-1)*-1
+    #
+    #         Examples
+    #     ---------
+    #     >>> import numpy as np
+    #     >>> from SeqMetrics import RegressionMetrics
+    #     >>> t = np.random.random(10)
+    #     >>> p = np.random.random(10)
+    #     >>> metrics= RegressionMetrics(t, p)
+    #     >>> metrics.log_nse()
+    #
+    #     """
+    #     return log_nse(true=self.true, predicted=self.predicted, epsilon=epsilon, treat_arrays=False)
 
     def log_prob(self) -> float:
         """
@@ -2769,91 +2769,91 @@ def spearmann_corr(
 
     return float(r)
 
-def spearmann_rank_corr(
-        true,
-        predicted,
-        treat_arrays: bool = True,
-        **treat_arrays_kws
-) -> float:
-    """Separmann rank correlation coefficient_.
+# def spearmann_rank_corr(
+#         true,
+#         predicted,
[…]
