+++ docs/source/quick_start.rst
@@ -64,7 +64,7 @@ ClassificationMetrics
 Streamlit App
 ==============
 The SeqMetrics library is available from the webapp which is deployed
-used stream [https://seqmetrics.streamlit.app/](https://seqmetrics.streamlit.app/)
+used streamlit app_ .
 
 You can also launch the app locally if you do not wish to use the web-based app. 
 Make sure you follow the below steps
@@ -77,11 +77,10 @@ Make sure you follow the below steps
 
 .. image:: imgs/fig2.jpg
     :align: center
-    :height: 600
-    :width: 400
+
 
 .. image:: imgs/fig3.jpg
     :align: center
-    :height: 600
-    :width: 400
 
+.. _app:
+    https://seqmetrics.streamlit.app/
