### 2024-05-12 b1b2bb1
do not repeat tests for class based api
