+++ tests/test_rgr.py
@@ -6,11 +6,8 @@ import site  # so that SeqMetrics directory is in path
 seqmet_dir = os.path.dirname(os.path.dirname(os.path.abspath(__file__)))
 site.addsitedir(seqmet_dir)
 
-import importlib
-
 import scipy
 import numpy as np
-import pandas as pd
 
 from scipy.stats import pearsonr, kendalltau, spearmanr
 
@@ -133,6 +130,7 @@ from SeqMetrics import log_cosh_error as sm_log_cosh_error
 from SeqMetrics import minkowski_distance as sm_minkowski_distance
 from SeqMetrics import tweedie_deviance_score as sm_tweedie_deviance_score
 from SeqMetrics import spearmann_corr
+from SeqMetrics import r2
 
 from SeqMetrics.utils import maybe_treat_arrays
 
@@ -326,44 +324,25 @@ class test_errors(unittest.TestCase):
         return
 
     def test_r2(self):
-        new_r2 = metrics.r2()
+        new_r2 = r2(t11, p11)
         _, _, rvalue, _, _ = scipy.stats.linregress(t11, p11)
         assert np.allclose(new_r2, rvalue ** 2)
 
-        new_r2 = metrics_large.r2()
+        new_r2 = r2(t_large, p_large)
         _, _, rvalue, _, _ = scipy.stats.linregress(t_large, p_large)
         assert np.allclose(new_r2, rvalue ** 2)
 
-        new_r2 = metrics_nan.r2()
+        new_r2 = r2(t_nan, p_nan)
         t_nan_, p_nan_ = maybe_treat_arrays(True, t_nan, p_nan, 'regression', remove_nan=True)
         _, _, rvalue, _, _ = scipy.stats.linregress(t_nan_, p_nan_)
         assert np.allclose(new_r2, rvalue ** 2)
 
-        new_r2 = metrics_neg.r2()
+        new_r2 = r2(t_neg, p_neg)
         _, _, rvalue, _, _ = scipy.stats.linregress(t_neg, p_neg)
         assert np.allclose(new_r2, rvalue ** 2)
         return
 
-    def test_mse_cls(self):
-        new_mse = metrics.mse()
-        sk_mse = mean_squared_error(t11, p11)
-        assert np.allclose(new_mse, sk_mse)
-
-        new_mse = metrics_large.mse()
-        sk_mse = mean_squared_error(t_large, p_large)
-        assert np.allclose(new_mse, sk_mse)
-
-        new_mse = metrics_nan.mse()
-        t_nan_, p_nan_ = maybe_treat_arrays(True, t_nan, p_nan, 'regression', remove_nan=True)
-        sk_mse = mean_squared_error(t_nan_, p_nan_)
-        assert np.allclose(new_mse, sk_mse)
-
-        new_mse = metrics_neg.mse()
-        sk_mse = mean_squared_error(t_neg, p_neg)
-        assert np.allclose(new_mse, sk_mse)
-        return
-
-    def test_mse_func(self):
+    def test_mse(self):
         new_mse = sm_mse(t11, p11)
         sk_mse = mean_squared_error(t11, p11)
         assert np.allclose(new_mse, sk_mse)
@@ -467,25 +446,6 @@ class test_errors(unittest.TestCase):
         new_nse_bound = nse_bound(t_neg, p_neg)
         assert np.allclose(new_nse_bound, -0.29295000460088483)
 
-        return
-
-    def test_r2_score_cls(self):
-        new_r2_score = metrics.r2_score()
-        sk_r2_score = r2_score(t11, p11)
-        assert np.allclose(new_r2_score, sk_r2_score)
-
-        new_r2_score = metrics_large.r2_score()
[…]
