+++ paper/paper.bib
@@ -57,3 +57,38 @@ year = {2018}
   year={2011},
   publisher={JMLR. org}
 }
+
+
+@article{Kratzert2022,
+    doi = {10.21105/joss.04050},
+    url = {https://doi.org/10.21105/joss.04050},
+    year = {2022},
+    publisher = {The Open Journal},
+    volume = {7},
+    number = {71},
+    pages = {4050},
+    author = {Frederik Kratzert and Martin Gauch and Grey Nearing and Daniel Klotz},
+    title = {NeuralHydrology --- A Python library for Deep Learning research in hydrology},
+    journal = {Journal of Open Source Software}
+ }
+
+
+ @misc{hydroeval2021,
+  title={hydroeval: an evaluator for streamflow time series in Python},
+  author={Hallouin, Thibault},
+  year={2021},
+  publisher={GitHub},
+  doi={10.5281/zenodo.2591217}
+}
+
+
+ @article{wade2018hydroerr,
+  title={Hydrostats: A Python Package for Characterizing Errors between Observed and Predicted Time Series},
+  author={Roberts, Wade and Williams, Gustavious and Jackson, Elise and Nelson, James and Ames, Daniel},
+  year={2021},
+  publisher={MDPI},
+  Journal = {Hydrology},
+  volume = {5},
+  doi={10.3390/hydrology5040066},
+  url = {https://www.mdpi.com/2306-5338/5/4/66}
+}
