### 2024-05-12 d2bb232
added lower limit of scipy and numpy
