+++ requirements.txt
@@ -1,2 +1,2 @@
-numpy<=1.26.4
-scipy<=1.13
+numpy<=1.26.4, >=1.17
+scipy<=1.13, >=1.4
