+++ docs/source/utils.rst
@@ -7,3 +7,9 @@ Utils
 .. autoclass:: SeqMetrics.utils.plot_metrics
    :members:
    :undoc-members:
+
+.. image:: imgs/reg1.png
+    :align: left
+
+.. image:: imgs/reg2.png
+    :align: right
