### 2024-05-12 #6450 @AtrCheema
@mstimberg We are almost done with the review. I apologize that the review quite some time which was not anticipated at the start.

I would first like to respond to the comments made by @SkafteNicki which are the most comprehensive one. Moreover they are endorsed and overlapped by the other two reviewers.

Moreover, I would like to thank all the three reviewers for taking the time to review the repository in detail. By addressing the comments, we have not only improved the overall quality of the package but also removed some bugs.

## Comments by @SkafteNicki

> You mention that [easy_mpl](https://easy-mpl.readthedocs.io/) is needed for plotting the metrics. However, it is not mentioned in the documentation or README that you can actually install this by writing pip install SeqMetrics[all]. Please add these additional install instructions.

**Response**: We have updated the readme and and documentation to add additional install instructions.

> Documentation for the class based API is essentially missing: https://seqmetrics.readthedocs.io/en/latest/rgr.html#SeqMetrics.RegressionMetrics. I know this is because it is just calling the functional API but it would then be great if there was a reference per metric to there functional counterpart.

**Response**: Initially we thought, adding the same documentation for methods of class based API would involve significant duplication. However, we have now added he documentation for class-based API as well for both [regression](https://seqmetrics.readthedocs.io/en/latest/rgr.html#class-based-api) and [classification](https://seqmetrics.readthedocs.io/en/latest/cls.html#class-based-api)

> Additionally, in the README.md of the project there are multiple related projects mentioned at the bottom that are not included in the paper ( [forecasting_metrics](https://gist.github.com/bshishov/5dc237f59f019b26145648e2124ca1c9), [hydroeval](https://github.com/ThibHlln/hydroeval) etc). I would like to ask the authors why these are not reference in the paper.

**Response**: The updated paper now contains reference to most of the packages listed in README.md.

> On the other hand, all the frameworks mentioned in the paper are not listed in the related section on the README.md. Again, minor stuff.

**Response**: The [updated README.md](https://github.com/AtrCheema/SeqMetrics?tab=readme-ov-file#related) now contains all the frameworks which are mentioned in the paper.

> The app should be better documented, especially for instructions for typing/pasting values. From the code I can see that a comma separated list is expected, but this is not clear from the instructions. A simple numpy array does not work for example. Including [fig2](https://github.com/AtrCheema/SeqMetrics/blob/master/paper/fig2.jpg) and [fig3](https://github.com/AtrCheema/SeqMetrics/blob/master/paper/fig3.jpg) to the documentation and README file would definitely help.

**Response**: The app can be used by typing/pasting data which is either comma separated or space separated. We have updated the  instructions on the app. Furthermore, the two figures are also added to README file and to documentation.

> Since it is a simple streamlit app that users can deploy themself without too much hassle, I really think the authors should consider adding instructions on how the app can be deployed by users locally (lets say that I do not trust streamlit servers with my data but still want the nice interface). This probably requires a bit of refactoring of the repository to include the app in the src directory and the addition of pip install SegMetrics[app] option for installing. Additionally, the paper should be updated to reflect that the webinterface can be self hosted.

**Response**: Launching the streamlit app locally requires installing the requirements including the streamlit package and then launching the streamlit app. We have explicitly added these steps in both [README](https://github.com/AtrCheema/SeqMetrics?tab=readme-ov-file#web-app) and [documentation](https://seqmetrics.readthedocs.io/en/latest/quick_start.html#streamlit-app).

> Here two different metrics are tested for a single given input. However, it is not clear at all why these tests are actually checking that implementation is correct.

**Response**: All the unit tests are now run for multiple inputs i.e. small values, large values (>e7), values with nans and -ve values. For all of these cases, the results are compared against a standard/reference. These standard/reference are more elaborated in response to next comment.

> (Important) Implement unit testing against other frameworks whenever possible. The authors are already doing this for certain classification metrics ([SeqMetrics/tests/test_cls.py](https://github.com/AtrCheema/SeqMetrics/blob/f1b885809910f222c4fabd337c31b485e43d0f81/tests/test_cls.py#L11-L12) Lines 11 to 12 in [f1b8858](https://github.com/AtrCheema/SeqMetrics/commit/f1b885809910f222c4fabd337c31b485e43d0f81)

    from sklearn.metrics import accuracy_score, precision_score, recall_score, f1_score
    from sklearn.metrics import confusion_matrix, balanced_accuracy_score
> ), however it should be done for most metrics. Metrics where there is no reference implementation to compare against should be tested for multiple values and preferably it should be more clear where the reference values comes from.

**Response**: We have modified the unit tests to include references. Overall, the references and corresponding metrics can be categorized into five groups.
1) The metrics where an implementation in a standard library is available. This included metrics from sklearn and Torchmetrics. All the metrics in SeqMetrics which are also available in these libraries, are compared with corresponding functions of these libraries.
2) The metrics for which an implementation is available in other libraries such as HydroErr, NeuralHydrology, skill metrics etc. Since enlisting these libraries in the requirements for tests would increase the number of requirements thereby making future development of SeqMetrics difficult. This is especially for the case where those libraries are no longer maintained. We have, however installed these libraries on a [colab notebook](https://github.com/AtrCheema/SeqMetrics/blob/master/tests/reference_for_seqmetric_tests.ipynb), calculated the reference values and then compared the SeqMetrics against these reference values in the tests. The colab notebook is provided in the tests for references.
3) The metrics whose calculation was too obvious such as ``std_ratio`` or ``gmean_diff``. For these metrics, a reference is not provided but their documentation is improved.
5) The metrics for which reference implementation was not available in a python library/package. However their code is available in the form of stackoverflow answers or github gist. For these metrics, we have copied the code (with reference) in the [colab notebook](https://github.com/AtrCheema/SeqMetrics/blob/master/tests/reference_for_seqmetric_tests.ipynb), and calculated the reference value. The tests are then run against these reference values.
4) The metrics for which no reference implementation was available in python. For these metrics we have provided the references for the formulas or the reference for the implementation in another language.
5) We also encountered two metrics for which we could not find any reference. We have removed these metrics for the time being.

> (Important) Currently only Python 3.7 is tested, which is officially [end-of-life](https://devguide.python.org/versions/). Either run CI that checks multiple versions of Python or at the very least a newer supported version

**Response**: We are now testing against [3.7 and 3.12](https://github.com/AtrCheema/SeqMetrics/blob/master/.github/workflows/py37.yml) which are the lower and upper python versions supported by this library.

> Test only run on ubuntu right now and no other major OS. I recommend that the authors either add test for other OS or explicitly state what OS are supported in there README.md

**Response**: We are now testing the library on [windows, ubuntu](https://github.com/AtrCheema/SeqMetrics/blob/master/.github/workflows/py37.yml) and [mac](https://github.com/AtrCheema/SeqMetrics/blob/master/.github/workflows/pymac.yml) with python 3.7 and python 3.12 versions.

> (Important) Because the CI only uses Python 3.7 the actual numpy version being tested is numpy-1.21.6, which is around 2 years old at this point. I see this as a overall consequence that the authors have not included some upper/lower bounds on supported numpy/scipy versions in the [requirements](https://github.com/AtrCheema/SeqMetrics/blob/master/requirements.txt) file.

**Response**: We are now testing the library for numpy [1.26.4](https://github.com/AtrCheema/SeqMetrics/blob/master/.github/workflows/np264.yml) and [1.17](https://github.com/AtrCheema/SeqMetrics/blob/master/.github/workflows/np116.yml) which are the upper and lower numpy versions supported by the library. The [setup](https://github.com/AtrCheema/SeqMetrics/blob/master/setup.py) and [requirement](https://github.com/AtrCheema/SeqMetrics/blob/master/requirements.txt) files are also updated to reflect this change.

> (Important) Missing community guidelines: are contributions welcome? what should a contribution look like etc.?

**Response**: We have added a [CONTRIBUTING.rst](https://github.com/AtrCheema/SeqMetrics/blob/master/CONTRIBUTING.rst) file, highlighting protocol for the potential contributors.

> (nitpicking) In [fig1](https://github.com/AtrCheema/SeqMetrics/blob/master/paper/fig1.jpg) of the paper the authors it is redundant to say "class-based api" both at the bottom and top of the figure (same goes for functional). Only mentioning this once should be enough?

**Response**: We have modifed the [figure 1](https://github.com/AtrCheema/SeqMetrics/blob/master/paper/fig1.jpg) by removing the "class-based api" at the bottom.

> (nitpicking) The overall resolution (pixel-wise) of figures in the paper is on the lower side and could be increase to help the readability of the text in the figures

**Response**: We have added the figures with higher resolution (900 dpi).
