+++ paper/paper.md
@@ -54,7 +54,7 @@ For instance, the metrics sub-module from Keras [@chollet2015keras] contains onl
 while scikit-learn's [@pedregosa2011scikit] metrics module covers 45 metrics. Torchmetrics library, [@detlefsen2022torchmetrics] 
 offers over 100 metrics. However, it provides only 48 which are intended
 for 1-dimensional numerical data. There are also some subject-specific
-libraries such as NeuralHydrology [Kratzert2022], hydroeval [hydroeval2021] and HydroErr [wade2018hydroerr], however these resources
+libraries such as NeuralHydrology [@Kratzert2022], hydroeval [@hydroeval2021] and HydroErr [@wade2018hydroerr], however these resources
 lack more general set of metrics which the experts in these demoains often require. 
 Addressing this gap is imperative to provide 
 researchers a unified platform for 
@@ -124,11 +124,15 @@ concerns or those without internet connection. The steps to launch streamlit GUI
 locally involve cloning the respository, isntalling the requirements and streamlit
 python package and then launching the streamlit app. These steps are given below
 
+```markdown
+
     git clone https://github.com/AtrCheema/SeqMetrics.git
     cd SeqMetrics
     pip install requirements.txt
     pip install streamlit
     streamlit run app.py
+
+```
 
 # Testing and documentation
 Following the ‘unit test’ protocol the library undergoes comprehensive testing 
