+++ CONTRIBUTING.rst
@@ -8,4 +8,7 @@ The contributions are welcome in the form of
 - Documentation improvements
 - New features
 
+If you add a new feature performance metric, you must add it for both the functional and class-based api, then write
+the tests with multiple inputs for the metric.
+
 The contributors must run the tests and ensure that they pass before submitting a pull request.
