### 2024-05-14 cdb058f
make scipy as optional dependency only
