+++ app.py
@@ -53,7 +53,7 @@ NAME_REG = {'calculate_all' : 'Calculate All Regression Metrics',
             'kge_bound': 'Bounded Kling-Gupta Efficiency',
             'kge_mod': 'Modified Kling-Gupta Efficiency',
             'kge_np': 'Non parametric Kling-Gupta Efficiency',
-            'kendaull_tau': 'Kendall’s tau',
+            'kendall_tau': 'Kendall’s tau',
             'kgeprime_bound': 'Bounded Version of the Modified Kling-Gupta Efficiency',
             'kgenp_bound': 'Bounded Version of the Non-Parametric Kling-Gupta Efficiency',
             'kl_sym': 'Symmetric Kullback-leibler Divergence',
