+++ CONTRIBUTING.rst
@@ -9,6 +9,7 @@ The contributions are welcome in the form of
 - New features
 
 If you add a new feature performance metric, you must add it for both the functional and class-based api, then write
-the tests with multiple inputs for the metric.
+the tests with multiple inputs for the metric. Also make sure that you add its documentation under docs directory
+and also specify it's type (either ``min`` or ``max``) in utils.py and then add its name in app.py file.
 
-The contributors must run the tests and ensure that they pass before submitting a pull request.
+If you add a new feature, make sure to run the tests and ensure that they pass before submitting a pull request.
