+++ app.py
@@ -130,7 +130,9 @@ NAME_REG = {'calculate_all' : 'Calculate All Regression Metrics',
             'watt_m' : 'Watterson’s M. Refrence',
             'wmape' : 'Weighted Mean Absolute Percent Error',
             'mre': "Mean Relative Error",
-            "mape_for_peaks": "MAPE for peaks"
+            "mape_for_peaks": "MAPE for peaks",
+            "legates_coeff_eff": "Legates Coefficient of Efficiency",
+            "relative_error": "Relative Error",
             }
 
 NAME_CLS = {'accuracy' : 'Accuracy',
