+++ SeqMetrics/__init__.py
@@ -127,6 +127,8 @@ from ._rgr import tweedie_deviance_score
 from ._rgr import mre
 # from ._rgr import spearmann_rank_corr todo, find reference for the code
 from ._rgr import mape_for_peaks
+from ._rgr import legates_coeff_eff
+from ._rgr import relative_error
 
 
 from ._cls import f1_score
