### 2024-05-14 0babc7f
remove relative error since it does not return float
