+++ SeqMetrics/__init__.py
@@ -129,6 +129,7 @@ from ._rgr import mre
 from ._rgr import mape_for_peaks
 from ._rgr import legates_coeff_eff
 # from ._rgr import relative_error
+from ._rgr import  manhattan_distance
 
 
 from ._cls import f1_score
