### 2024-05-14 #6450 @AtrCheema
@mstimberg yes correct. Indeed, most of the comments by other two reviewers have also been addressed while addressing @SkafteNicki 's comments. However, two comments (adding reference equation for each metric, and a more detailed discussion of current literature/libraries on performance metrics) are remaining. We are currently improving the docs and I will respond to both reviewers as soon as we are done.
