### 2024-05-17 6092490
Adding equations in regression metrics functional APIs
