+++ SeqMetrics/_rgr.py
@@ -2350,6 +2350,9 @@ def r2(true, predicted, treat_arrays: bool = True,
     .. _explains:
         https://data.library.virginia.edu/is-r-squared-useless/
 
+    .. math ::
+        R^2 = \\left( \\frac{\\sum_{i=1}^{N} \\left( \\frac{true_i - \\bar{true}}{\\sigma_{true}} \\cdot \\frac{predicted_i - \\bar{predicted}}{\\sigma_{predicted}} \\right)}{N - 1} \\right)^2
+
     Parameters
     ----------
     true :
@@ -2395,6 +2398,9 @@ def nse(true, predicted, treat_arrays: bool = True,
         Performance measures and evaluation criteria. Transactions of the ASABE, 58(6), 1763-1785.
     - Krause, P., Boyle, D., & Bäse, F. (2005). Comparison of different efficiency criteria for hydrological
         model assessment. Adv. Geosci., 5, 89-97. https://dx.doi.org/10.5194/adgeo-5-89-2005.
+
+    .. math::
+        \\text{NSE} = 1 - \\frac{\\sum_{i=1}^{N} (predicted_i - true_i)^2}{\\sum_{i=1}^{N} (true_i - \\bar{true})^2}
     Parameters
     ----------
     true :
@@ -2424,6 +2430,10 @@ def nse_alpha(true, predicted, treat_arrays: bool = True,
     """
     Alpha decomposition of the NSE, see `Gupta_ et al. 2009 <https://doi.org/10.1029/97WR03495>`_
     used in `kratzert et al., 2018 <>`_
+
+    .. math::
+        \\text{NSE}_{\\text{alpha}} = \\frac{\\sigma_{\\text{predicted}}}{\\sigma_{\\text{true}}}
+
 
     Returns
     -------
@@ -2460,6 +2470,9 @@ def nse_beta(true, predicted, treat_arrays: bool = True,
     .. _Gupta:
         https://doi.org/10.1016/j.jhydrol.2009.08.003
 
+    .. math::
+        \\text{NSE}_{\\text{beta}} = \\frac{\\mu_{\\text{predicted}} - \\mu_{\\text{true}}}{\\sigma_{\\text{true}}}
+
     Returns
     -------
     float
@@ -2495,6 +2508,9 @@ def nse_mod(true, predicted, treat_arrays: bool = True,
     Gives less weightage to outliers if j=1 and if j>1 then it gives more
     weightage to outliers. Reference: Krause et al., 2005
 
+    .. math::
+        \\text{NSE}_{\\text{mod}} = 1 - \\frac{\\sum_{i=1}^{N} \\left| \\text{predicted}_i - \\text{true}_i \\right|^j}{\\sum_{i=1}^{N} \\left| \\text{true}_i - \\bar{\text{true}} \\right|^j}
+
     Parameters
     ----------
     true :
@@ -2528,6 +2544,10 @@ def nse_rel(true, predicted,
     """
     Relative NSE.
 
+    .. math::
+        \\text{NSE}_{\\text{rel}} = 1 - \\frac{\\sum_{i=1}^{N} \\left( \\frac{|\\text{predicted}_i - \\text{true}_i|}{\\text{true}_i} \\right)^2}{\\sum_{i=1}^{N} \\left( \\frac{|\\text{true}_i - \\overline{\\text{true}}|}{\\overline{\\text{true}}} \\right)^2}
+
+
     Parameters
     ----------
     true :
@@ -2558,8 +2578,13 @@ def nse_bound(true, predicted, treat_arrays: bool = True,
     """
     Bounded Version of the Nash-Sutcliffe Efficiency (nse_)
 
+    .. math::
[…]
