### 2024-05-22 #6450 @FATelarico
> @mstimberg yes correct. Indeed, most of the comments by other two reviewers have also been addressed while addressing @SkafteNicki 's comments. However, two comments (adding reference equation for each metric, and a more detailed discussion of current literature/libraries on performance metrics) are remaining. We are currently improving the docs and I will respond to both reviewers as soon as we are done.

Thank you for the update! I understand the remaining aspects, mainly those pertaining to my review specifically ([above](https://github.com/openjournals/joss-reviews/issues/6450#issuecomment-1985733025)), will be dealt with soon. However, I wanted to take the time to congratulate the authors on the great job in updating the dependencies and improving the documentation. I managed to install the library on both Ubuntu 22 and Winows 11 without downgrading Python or any other module. All the main functionalities are as expected and documented.

I am looking forward to a new version of the paper that engages more with the literature (incl. having more references) and is finally polished for publication.
