### 2024-05-25 9bf3355
Adding formulas for all regression metrics functional APIs
