+++ SeqMetrics/_rgr.py
@@ -2714,7 +2714,7 @@ def kge(true,
     .. math::
         \\beta = \\frac{\\mu_{\\text{predicted}}}{\\mu_{\\text{true}}}
 
-
+    10.5194/essd-12-2043-2020
 
     output:
         If return_all is True, it returns a numpy array of shape (4, ) containing
@@ -2897,6 +2897,9 @@ def spearmann_corr(
 )->float:
     """
     Spearman correlation coefficient
+
+    .. math::
+        r = \\frac{\\sum_{i=1}^{n} \\left( R_{t,i} - \\overline{R_t} \\right) \\left( R_{p,i} - \\overline{R_p} \\right)}{\\sqrt{ \\sum_{i=1}^{n} \\left( R_{t,i} - \\overline{R_t} \\right)^2 \\sum_{i=1}^{n} \\left( R_{p,i} - \\overline{R_p} \\right)^2 }}
 
     Parameters
     ----------
@@ -4068,6 +4071,11 @@ def cronbach_alpha(true, predicted, treat_arrays: bool = True,
     It is a measure of internal consitency of data. See ucla and stackoverflow_
     pages for more info.
 
+    .. math::
+        alpha = \\frac{N}{N - 1} \\left(1 - \\frac{\\sum_{i=1}^{N} \\sigma^2_{i}}{\\sigma^2_{\\text{total}}}\\right)
+
+    https://doi.org/10.1016/B0-12-369398-5/00396-0
+
     .. _ucla:
         https://stats.idre.ucla.edu/spss/faq/what-does-cronbachs-alpha-mean/
 
@@ -4110,6 +4118,10 @@ def centered_rms_dev(true, predicted, treat_arrays: bool = True,
     where p is the predicted values, r is the true values, and
     N is the total number of values in p & r.
 
+    .. math::
+        CRMSD = \\sqrt{\\frac{1}{N} \\sum_{i=1}^{N} \\left( (p_i - \\text{mean}(p)) - (r_i - \\text{mean}(r)) \\right)^2}
+
+    https://doi.org/10.1016/B0-12-227410-5/00612-8
     Output:
     CRMSDIFF : centered root-mean-square (RMS) difference (E')^2
 
@@ -4151,6 +4163,11 @@ def cosine_similarity(true, predicted, treat_arrays: bool = True,
     the same orientation have a cosine similarity of 1, two vectors oriented
     at 90° relative to each other have a similarity of 0, and two vectors diametrically
     opposed have a similarity of -1, independent of their magnitude. See_
+
+    .. math::
+        \\text{Cosine Similarity} = \\frac{\\sum_{i=1}^{n} \\text{true}_i \\cdot \\text{predicted}_i}{\\sqrt{\\sum_{i=1}^{n} (\\text{true}_i)^2} \\cdot \\sqrt{\\sum_{i=1}^{n} (\\text{predicted}_i)^2}}
+
+    https://doi.org/10.1016/B978-0-12-804452-0.00002-6
 
     .. _see:
         https://en.wikipedia.org/wiki/Cosine_similarity
@@ -4230,6 +4247,11 @@ def euclid_distance(true, predicted, treat_arrays: bool = True,
                     **treat_arrays_kws) -> float:
     """Euclidian distance
 
+    .. math::
+        D = \\sqrt{\\sum_{i=1}^{n} (\\text{true}_i - \\text{predicted}_i)^2}
+
+    https://doi.org/10.1016/B978-0-12-088735-4.50006-7
+
     Referneces: Kennard et al., 2010
     Parameters
     ----------
@@ -4258,6 +4280,10 @@ def exp_var_score(true, predicted, treat_arrays: bool = True, weights=None,
     """
[…]
