### 2024-05-25 61c9fc8
Adding formulas for classification Metrics functional API
