+++ SeqMetrics/_cls.py
@@ -677,6 +677,8 @@ def cross_entropy(true, predicted, epsilon=1e-12) -> float:
     Computes cross entropy between targets (encoded as one-hot vectors)
     and predictions.
 
+    .. math::
+        CE = - \\frac{1}{N} \\sum_{i=1}^{N} \\left[ y_i \\log(\\hat{y}_i + \\epsilon) \\right]
     Returns
     -------
     scalar
@@ -720,6 +722,9 @@ def accuracy(true, predicted, normalize: bool = True) -> float:
     """
     calculates accuracy
 
+    .. math::
+        \\text{Accuracy} = \\frac{\\sum_{i=1}^{N} \\mathbb{1}(true_i = predicted_i)}{N}
+
     Parameters
     ----------
     normalize : bool
@@ -753,6 +758,15 @@ def precision(true, predicted, average=None):
     It is number of correct positive predictions divided by the total
     number of positive predictions.
     TP/(TP+FP)
+
+    .. math::
+        \\text{Precision}_{\\text{micro}} = \\frac{\\sum TP}{\\sum (TP + FP)}
+
+    .. math::
+        \\text{Precision}_{\\text{macro}} = \\frac{1}{N} \\sum_{i=1}^{N} \\frac{TP_i}{TP_i + FP_i}
+
+    .. math::
+        \\text{Precision}_{\\text{weighted}} = \\frac{\\sum_{i=1}^{N} (TP_i + FN_i) \\cdot \\frac{TP_i}{TP_i + FP_i}}{\\sum_{i=1}^{N} (TP_i + FN_i)}
 
     Parameters
     ----------
@@ -799,8 +813,18 @@ def recall(true, predicted, average=None):
     """
     It is also called sensitivity or true positive rate. It is
     number of correct positive predictions divided by the total number of positives
-    Formula :
-        True Posivitive / True Positive + False Negative
+
+    .. math::
+        \\text{Recall} = \\frac{\\text{True Positive}}{\\text{True Positive} + \\text{False Negative}}
+
+    .. math::
+        \\text{Recall}_{\\text{micro}} = \\frac{\\sum_{i=1}^{n} \\text{TP}_i}{\\sum_{i=1}^{n} (\\text{TP}_i + \\text{FN}_i)}
+
+    .. math::
+        \\text{Recall}_{\\text{macro}} = \\frac{1}{n} \\sum_{i=1}^{n} \\frac{\\text{TP}_i}{\\text{TP}_i + \\text{FN}_i}
+
+    .. math::
+        \\text{Recall}_{\\text{weighted}} = \\sum_{i=1}^{n} w_i \\cdot \\frac{\\text{TP}_i}{\\text{TP}_i + \\text{FN}_i}
 
     Parameters
     ----------
@@ -849,8 +873,8 @@ def specificity(true, predicted, average=None):
     the predictions are negative when the true labels are also negative.
     It is number of correct negative predictions divided by the total number of negatives.
 
-    Formula :
-    TN / TN+FP
+    .. math::
+        \\text{Specificity} = \\frac{TN}{TN + FP}
 
     Parameters
     ----------
@@ -893,6 +917,9 @@ def balanced_accuracy(true, predicted, average=None) -> float:
     balanced accuracy.
     It performs better on imbalanced datasets.
 
+    .. math::
+        \\text{Balanced Accuracy} = \\frac{1}{C} \\sum_{i=1}^{C} \\frac{TP_i}{TP_i + FN_i}
+
     Parameters
     ----------
     true : ture/observed/actual/target values. It must be a numpy array,
@@ -919,28 +946,30 @@ def balanced_accuracy(true, predicted, average=None) -> float:
 
[…]
