### 2024-05-28 #6450 @SkafteNicki
Hi @mstimberg and @AtrCheema, sorry for the late reply from my side, I been sick the last week.
I am overall really happy with the changes the authors have made to the code. I quickly looked through it and it is of a much higher quality now. I also looked through the paper and would say that is also good now.
I will go ahead and update my checklist.
