### 2024-05-28 b5d83c9
Adding formulas for classification metrics Class based API
