+++ SeqMetrics/_cls.py
@@ -217,6 +217,9 @@ class ClassificationMetrics(Metrics):
         """
         calculates accuracy
 
+        .. math::
+            \\text{Accuracy} = \\frac{\\sum_{i=1}^{N} \\mathbb{1}(true_i = predicted_i)}{N}
+
         Examples
         --------
         >>> import numpy as np
@@ -313,7 +316,14 @@ class ClassificationMetrics(Metrics):
         Returns precision score, also called positive predictive value.
         It is number of correct positive predictions divided by the total
         number of positive predictions.
-        TP/(TP+FP)
+        .. math::
+            \\text{Precision}_{\\text{micro}} = \\frac{\\sum TP}{\\sum (TP + FP)}
+
+        .. math::
+            \\text{Precision}_{\\text{macro}} = \\frac{1}{N} \\sum_{i=1}^{N} \\frac{TP_i}{TP_i + FP_i}
+
+        .. math::
+            \\text{Precision}_{\\text{weighted}} = \\frac{\\sum_{i=1}^{N} (TP_i + FN_i) \\cdot \\frac{TP_i}{TP_i + FP_i}}{\\sum_{i=1}^{N} (TP_i + FN_i)}
 
         Examples
         --------
@@ -333,8 +343,18 @@ class ClassificationMetrics(Metrics):
         """
         It is also called sensitivity or true positive rate. It is
         number of correct positive predictions divided by the total number of positives
-        Formula :
-            True Posivitive / True Positive + False Negative
+
+        .. math::
+            \\text{Recall} = \\frac{\\text{True Positive}}{\\text{True Positive} + \\text{False Negative}}
+
+        .. math::
+            \\text{Recall}_{\\text{micro}} = \\frac{\\sum_{i=1}^{n} \\text{TP}_i}{\\sum_{i=1}^{n} (\\text{TP}_i + \\text{FN}_i)}
+
+        .. math::
+            \\text{Recall}_{\\text{macro}} = \\frac{1}{n} \\sum_{i=1}^{n} \\frac{\\text{TP}_i}{\\text{TP}_i + \\text{FN}_i}
+
+        .. math::
+            \\text{Recall}_{\\text{weighted}} = \\sum_{i=1}^{n} w_i \\cdot \\frac{\\text{TP}_i}{\\text{TP}_i + \\text{FN}_i}
 
         Examples
         --------
@@ -354,8 +374,8 @@ class ClassificationMetrics(Metrics):
         the predictions are negative when the true labels are also negative.
         It is number of correct negative predictions divided by the total number of negatives.
 
-        Formula :
-        TN / TN+FP
+        .. math::
+            \\text{Specificity} = \\frac{TN}{TN + FP}
 
         Examples
         --------
@@ -375,6 +395,9 @@ class ClassificationMetrics(Metrics):
         """
         balanced accuracy.
         It performs better on imbalanced datasets.
+
+        .. math::
+            \\text{Balanced Accuracy} = \\frac{1}{C} \\sum_{i=1}^{C} \\frac{TP_i}{TP_i + FN_i}
 
         Examples
         --------
@@ -414,20 +437,23 @@ class ClassificationMetrics(Metrics):
 
     def f1_score(self, average=None) -> Union[np.ndarray, float]:
         """
-               Calculates f1 score according to following formula
-               f1_score = 2 * (precision * recall)  / (precision + recall)
+           Calculates f1 score according to following formula
+           f1_score = 2 * (precision * recall)  / (precision + recall)
 
-               Examples
-               --------
-               >>> import numpy as np
[…]
