+++ .readthedocs.yml
@@ -13,9 +13,9 @@ build:
   tools:
     python: "3.8"
 
-# If using Sphinx, optionally build your docs in additional formats such as PDF
-formats:
-    - pdf
+## If using Sphinx, optionally build your docs in additional formats such as PDF
+#formats:
+#    - pdf
 
 # Optionally declare the Python requirements required to build your docs
 python:
