### 2024-05-28 9abb8b5
Adding formulas for Regression Metrics Class based API
