+++ SeqMetrics/_rgr.py
@@ -107,6 +107,9 @@ class RegressionMetrics(Metrics):
         """Anomaly correction coefficient. See Langland_ et al., 2012; Miyakoda_ et al., 1972
         and Murphy_ et al., 1989.
 
+        .. math::
+            ACC = \\frac{\\sum_{i=1}^{N} \\left( (\\text{predicted}_i - \\overline{\\text{predicted}})(\\text{true}_i - \\overline{\\text{true}}) \\right)}{(N-1) \\cdot \\sigma_{\\text{true}} \\cdot \\sigma_{\\text{predicted}}}
+
         .. _Langland:
             https://doi.org/10.3402/tellusa.v64i0.17531
 
@@ -128,6 +131,9 @@ class RegressionMetrics(Metrics):
 
     def adjusted_r2(self) -> float:
         """Adjusted R squared.
+
+        .. math::
+            \\text{Adjusted } R^2 = 1 - \\left( \\frac{(1 - R^2) \\cdot (n - 1)}{n - k - 1} \\right)
 
         Examples
         ---------
@@ -178,6 +184,9 @@ class RegressionMetrics(Metrics):
         """
         Akaike_ Information Criterion. Modifying from this source_
 
+        .. math::
+            AIC = n \\cdot \\ln\\left(\\frac{\\sum_{i=1}^{n} (\\text{true}_i - \\text{predicted}_i)^2}{n}\\right) + 2p
+
         .. _Akaike:
             https://doi.org/10.1007/978-1-4612-1694-0_15
 
@@ -198,6 +207,9 @@ class RegressionMetrics(Metrics):
     def aitchison(self, center='mean') -> float:
         """Aitchison distance. used in Zhang_ et al., 2020
 
+        .. math::
+            d_{\\text{Aitchison}} = \\sqrt{\\sum_{i=1}^{n} \\left( \\log(\\text{true}_i) - \\text{center}(\\log(\\text{true})) - \\left(\\log(\\text{predicted}_i) - \\text{center}(\\log(\\text{predicted}))\\right) \\right)^2}
+
         .. _Zhang:
             https://doi.org/10.5194/hess-24-2505-2020
 
@@ -216,6 +228,9 @@ class RegressionMetrics(Metrics):
     def amemiya_adj_r2(self) -> float:
         """Amemiya's Adjusted R-squared
 
+        .. math::
+            R^2_{\\text{adj, Amemiya}} = 1 - \\left( \\frac{(1 - R^2) \\cdot (n + k)}{n - k - 1} \\right)
+
         Examples
         ---------
         >>> import numpy as np
@@ -229,6 +244,9 @@ class RegressionMetrics(Metrics):
 
     def amemiya_pred_criterion(self) -> float:
         """Amemiya's Prediction Criterion
+
+        .. math::
+            \\text{APC} = \\left( \\frac{n + k}{n - k} \\right) \\left( \\frac{1}{n} \\sum_{i=1}^{n} (\\text{true}_i - \\text{predicted}_i)^2 \\right)
 
         Examples
         ---------
@@ -271,6 +289,9 @@ class RegressionMetrics(Metrics):
         model chosen by the BIC is either the same as that chosen by the AIC, or one
         with fewer terms. This is because the BIC penalises the number of parameters
         more heavily than the AIC.  Modified after RegscorePy_.
+
+        .. math::
+            BIC = n \\cdot \\ln\\left(\\frac{\\text{SSE}}{n}\\right) + p \\cdot \\ln(n)
 
         .. _BIC:
[…]
