### 2024-05-29 1c9882b
Updating mape_for_peaks
