+++ SeqMetrics/_rgr.py
@@ -7615,8 +7615,8 @@ def mape_for_peaks(
     if peaks.size == 0:
         return np.nan
 
-    # true = true[peaks].values
-    # predicted = predicted[peaks].values
+    true = true[peaks]
+    predicted = predicted[peaks]
 
     return mape(true, predicted, treat_arrays=False)
 
