### 2024-05-30 a859f48
Adding some references
