### 2024-07-01 8d6cb0c
added deeper comparison with other libaries in paper
