+++ paper/paper.bib
@@ -92,3 +92,50 @@ year = {2018}
   doi={10.3390/hydrology5040066},
   url = {https://www.mdpi.com/2306-5338/5/4/66}
 }
+
+
+@inproceedings{Wang2020,
+author = {Wang, Ying and Wen, Ming and Liu, Yepang and Wang, Yibo and Li, Zhenming and Wang, Chao and Yu, Hai and Cheung, Shing-Chi and Xu, Chang and Zhu, Zhiliang},
+title = {Watchman: monitoring dependency conflicts for Python library ecosystem},
+year = {2020},
+isbn = {9781450371216},
+publisher = {Association for Computing Machinery},
+address = {New York, NY, USA},
+url = {https://doi.org/10.1145/3377811.3380426},
+doi = {10.1145/3377811.3380426},
+abstract = {The PyPI ecosystem has indexed millions of Python libraries to allow developers to automatically download and install dependencies of their projects based on the specified version constraints. Despite the convenience brought by automation, version constraints in Python projects can easily conflict, resulting in build failures. We refer to such conflicts as <u>D</u>ependency <u>C</u>onfict (DC) issues. Although DC issues are common in Python projects, developers lack tool support to gain a comprehensive knowledge for diagnosing the root causes of these issues. In this paper, we conducted an empirical study on 235 real-world DC issues. We studied the manifestation patterns and fixing strategies of these issues and found several key factors that can lead to DC issues and their regressions. Based on our findings, we designed and implemented Watchman, a technique to continuously monitor dependency conflicts for the PyPI ecosystem. In our evaluation, Watchman analyzed PyPI snapshots between 11 Jul 2019 and 16 Aug 2019, and found 117 potential DC issues. We reported these issues to the developers of the corresponding projects. So far, 63 issues have been confirmed, 38 of which have been quickly fixed by applying our suggested patches.},
+booktitle = {Proceedings of the ACM/IEEE 42nd International Conference on Software Engineering},
+pages = {125–135},
+numpages = {11},
+keywords = {Python, dependency conflicts, software ecosystem},
+location = {Seoul, South Korea},
+series = {ICSE '20}
+}
+
+@inproceedings{@Mukherjee2021,
+author = {Mukherjee, Suchita and Almanza, Abigail and Rubio-Gonz\'{a}lez, Cindy},
+title = {Fixing dependency errors for Python build reproducibility},
+year = {2021},
+isbn = {9781450384599},
+publisher = {Association for Computing Machinery},
+address = {New York, NY, USA},
+url = {https://doi.org/10.1145/3460319.3464797},
+doi = {10.1145/3460319.3464797},
[…]
