+++ docs/source/rgr.rst
@@ -127,6 +127,7 @@ metrics.
 .. autofunction:: SeqMetrics.mre
 .. autofunction:: SeqMetrics.mape_for_peaks
 .. autofunction:: SeqMetrics.legates_coeff_eff
+.. autofunction:: SeqMetrics.manhattan_distance
 
 
 .. autofunction:: SeqMetrics.calculate_hydro_metrics
