### 2024-07-01 #6450 @AtrCheema
Thank you everyone, especially the editor and handling editor to staying with us and your patience.

## Comments by @FATelarico

> As I have encountered significant issue with this sort of flaws in the past, I invite the authors to check the entire software on the newest stable release of Python 3 and update all the main packages (especially numpy, of which most people probably have a much more recent version).

**Response:** We have updated the gh workflows to test the package against python 3.7 and 3.12. We have also made sure that all the unit tests pass for numpy 1.17 to 1.26.4. The requirements files and setup.py file also reflect these changes.

> Personally, I had no problems running the programm on Python (downgraded to 3.7) under both Ubuntu 22. But it would not run on Windows 10 (64 bit) with Python 3.12.2 without downgrading. However, compatibility should be verified after the program is fully updated and the relevant information should be appended to the README.md file only then.

**Response:** We are now testing the code on ubuntu, windows and MacOS using gh workflows.

> I am not completely sure this is the right heading under which to put this comment, but it relates to 'claims' the paper makes. In fact, I did not see satisfactory indications of the tests' robustness. @SkafteNicki wrote extensively and better than I could about it

**Response:** As has been mentioned in response to @SkafteNicki comments, we have completely revised the tests.

> There are ready-made templates to add community guidelines. Consider just copy-pasting them and adapting the content to your desires. For instance: https://bttger.github.io/contributing-gen-web/ which is based on [contributing-gen](https://github.com/bttger/contributing-gen)

**Response:** We are open to contributions and have added [contributing.rst](https://github.com/AtrCheema/SeqMetrics/blob/master/CONTRIBUTING.rst) file.

> Content referrable under this point is contained in rows 24-31 of the Statement of need section. I would like the authors to consider shortening these passages and add a separate heading explictly dedicated to comparing their sofware to Keras, scikit-learn, Torchmetrics, forecasting_metrics, hydroeval, and other. They do not need to necessarily consider all of them, but at least the most widely used. In particular, the paper would benefit from a clear description of (some of) the use-cases in which SeqMetrics is technically preferable to existing alternatives as opposed to application in which its main added value is the GUI. For instance, the emphasis here is clearly on tabular and time-series, one-dimension data. But reading the paper, at times, one may forget it.

**Response:** Thanks for the comment. We have improved the section 'statement of need' by enhancing more detailed comparison with other libraries. This include limitation due to their dependencies and complex usage. However, we have avoided some limitations of these libraries which we think maybe removed in future. For example, the following code will fail because torchmetrics library does not accept python list/numpy array /pandas DataFrame  as input.

```python
from  torchmetrics.functional.regression import r2_score
import numpy as np
import pandas as pd
r2_score(np.array([1,2,3]), np.array([1.1, 2.2, 3.3]))
r2_score([1,2,3], [1.1, 2.2, 3.3])
r2_score(pd.DataFrame(np.array([1,2,3])), pd.DataFrame(np.array([1.1, 2.2, 3.3])))
```

> Five references seem too few for a paper that should help SeqMetrics stand out in a rather crowded field. Even if abovementioned suggestions to include additional sections are rejected, this issue ought to be settled through a more intenst dialogue with existing tools.

**Response:** We have added more references and the total number of references are now 10.

## Comments by @y1my1

> It's great that the authors provide documentation through [readthedocs](https://seqmetrics.readthedocs.io/en/latest/). The authors put a lot of effort there providing important information about the computation of metrics, like the formulas of some metrics. However, for some of the metrics, the authors just provided a reference there; it would be great if the authors could at least provide a formula that helps users understand the underhood of the computation. And it would be beneficial if the authors could write a concise introduction there.

**Response:** We have improved the documentation and have provided the brief explanation, equation and reference for each performance metric in the documentation.
