### 2024-07-01 #6450 @FATelarico
@AtrCheema Thank you very much for the review. I am almost completely satisfied, this is really a great job!

I would just suggest the following (feel free to argue agains anything if I am mistaken):

-  One of the references (rows 102-104) is to an ArXiv preprint. But the paper's was published on the _Interdisciplinary Journal of Information, Knowledge, and Management_ ([doi: 10.28945/4184](https://doi.org/10.28945/4184))
- On row 15: 'minimal depency', but should be _depencies_?
- A typo on row 34 ('demoains' instead of _domains_)
- There are some '(?)' on rows 36-41. I guess they are missing citations. If so, consider respectively: [Detlefsen et al 2022](https://doi.org/10.21105/joss.04101); [Cao et al 2022](https://doi.org/10.1109/TSE.2022.3191353)
- Improve for clarity (rows 53-56) for example: _Regression metrics such as R-squared, Mean Squared Error (MSE), and Root MSE are provided for use with continuous variables. On the other hand, for categorical data, there are classification metrics including accuracy, precision, recall, F1 score, and the area under the ROC curve, among others._
- Typo: 'From programming perspective' instead of _From a programming perspective_ (row 57)
- Typo: 'isntalling' instead of _installing_ (row 89)
- Missing punctuation: 'Following the "unit test" protocol the library' instead of 'protocol_,_ the' (row 92)
