### 2024-07-01 bdd0f2d
minor improvements in paper
