+++ paper/paper.bib
@@ -16,16 +16,16 @@ year = {2008}
 @article{Botchkarev,
 author = {Alexei Botchkarev},
 title = {Performance metrics (error measures) in machine learning regression, forecasting and prognostics: Properties and typology},
-journal = {ArXiv},
-volume = {},
+journal = {Interdisciplinary Journal of Information, Knowledge, and Management},
+volume = {14},
 number = {},
-pages = {},
+pages = {45-76},
 keywords = {Performance metrics, model evaluation},
-doi = {10.48550/arXiv.1809.03006},
+doi = {10.28945/4184},
 url = {https://arxiv.org/abs/1809.03006},
 eprint = {},
 abstract = {Performance metrics (error measures) are vital components of the evaluation frameworks in various fields. The intention of this study was to overview of a variety of performance metrics and approaches to their classification. The main goal of the study was to develop a typology that will help to improve our knowledge and understanding of metrics and facilitate their selection in machine learning regression, forecasting and prognostics. Based on the analysis of the structure of numerous performance metrics, we propose a framework of metrics which includes four (4) categories: primary metrics, extended metrics, composite metrics, and hybrid sets of metrics. The paper identified three (3) key components (dimensions) that determine the structure and properties of primary metrics: method of determining point distance, method of normalization, method of aggregation of point distances over a data set.},
-year = {2018}
+year = {2019}
 }
 
 @misc{chollet2015keras,
@@ -112,7 +112,7 @@ location = {Seoul, South Korea},
 series = {ICSE '20}
 }
 
-@inproceedings{@Mukherjee2021,
+@inproceedings{Mukherjee2021,
 author = {Mukherjee, Suchita and Almanza, Abigail and Rubio-Gonz\'{a}lez, Cindy},
 title = {Fixing dependency errors for Python build reproducibility},
 year = {2021},
@@ -130,7 +130,7 @@ location = {Virtual, Denmark},
 series = {ISSTA 2021}
 }
 
-@article{@paszke2019pytorch,
+@article{paszke2019pytorch,
   title={Pytorch: An imperative style, high-performance deep learning library},
   author={Paszke, Adam and Gross, Sam and Massa, Francisco and Lerer, Adam and Bradbury, James and Chanan, Gregory and Killeen, Trevor and Lin, Zeming and Gimelshein, Natalia and Antiga, Luca and others},
   journal={Advances in neural information processing systems},
