### 2024-07-06 2b45517
updated some references
