+++ SeqMetrics/_rgr.py
@@ -572,13 +572,16 @@ class RegressionMetrics(Metrics):
     def fdc_fhv(self, h: float = 0.02) -> float:
         """
         modified Kratzert2018_ code. Peak flow bias of the flow duration curve (Yilmaz 2008).
-        used in kratzert et al., 2018
+        used in kratzert_ et al., 2019
 
         .. math::
             FHV = \\frac{\\sum_{i=1}^{k} (predicted_i - true_i)}{\\sum_{i=1}^{k} true_i} \\times 100
 
         .. _Kratzert2018:
             https://github.com/kratzert/ealstm_regional_modeling/blob/64a446e9012ecd601e0a9680246d3bbf3f002f6d/papercode/metrics.py#L190
+
+        .. _kratzert
+            https://hess.copernicus.org/articles/23/5089/2019/hess-23-5089-2019.html
 
         Examples
         ---------
@@ -594,13 +597,16 @@ class RegressionMetrics(Metrics):
     def fdc_flv(self, low_flow: float = 0.3) -> float:
         """
         bias of the bottom 30 % low flows. modified Kratzert_ code
-        used in kratzert et al., 2018
+        used in kratzert_ et al., 2019
 
         .. math::
             \\text{FLV} = -1 \\times \\frac{\\sum (\\log(\\text{predicted}) - \\min(\\log(\\text{predicted}))) - \\sum (\\log(\\text{true}) - \\min(\\log(\\text{true})))}{\\sum (\\log(\\text{true}) - \\min(\\log(\\text{true}))) + 1 \\times 10^{-6}}
 
         .. _Kratzert:
             https://github.com/kratzert/ealstm_regional_modeling/blob/64a446e9012ecd601e0a9680246d3bbf3f002f6d/papercode/metrics.py#L237
+
+        .. _kratzert
+            https://hess.copernicus.org/articles/23/5089/2019/hess-23-5089-2019.html
 
         Examples
         ---------
@@ -1670,11 +1676,17 @@ class RegressionMetrics(Metrics):
 
     def nse_alpha(self) -> float:
         """
-        Alpha decomposition of the NSE, see `Gupta_ et al. 2009 <https://doi.org/10.1029/97WR03495>`_
-        used in `kratzert et al., 2018 <>`_
+        Alpha decomposition of the NSE, see Gupta_ et al. 2009
+        used in kratzert_ et al., 2019 .
 
         .. math::
             \\text{NSE}_{\\text{alpha}} = \\frac{\\sigma_{\\text{predicted}}}{\\sigma_{\\text{true}}}
+
+        .. _Gupta:
+            https://doi.org/10.1016/j.jhydrol.2009.08.003
+
+        .. _kratzert:
+            https://hess.copernicus.org/articles/23/5089/2019/hess-23-5089-2019.html
 
         Examples
         ---------
@@ -1689,14 +1701,17 @@ class RegressionMetrics(Metrics):
 
     def nse_beta(self) -> float:
         """
-        Beta decomposition of NSE. See Gupta_ et al. 2009 .
-        used in kratzert et al., 2018
+        Beta decomposition of NSE. See Gupta_ et al. 2009
+        used in kratzert_ et al., 2019 .
 
         .. math::
             \\text{NSE}_{\\text{beta}} = \\frac{\\mu_{\\text{predicted}} - \\mu_{\\text{true}}}{\\sigma_{\\text{true}}}
 
         .. _Gupta:
             https://doi.org/10.1016/j.jhydrol.2009.08.003
+
+        .. _kratzert:
[…]
