### 2024-07-07 #6450 @FATelarico
> Dear @FATelarico Thanks for highlighted the problems. I have incorporated the changes you mentioned.

Excellent job @AtrCheema! I updates my [checklist](https://github.com/openjournals/joss-reviews/issues/6450#issuecomment-1985667559). You're good to go as far as I'm concerned! 🦾
