+++ SeqMetrics/_rgr.py
@@ -104,19 +104,12 @@ class RegressionMetrics(Metrics):
     #     return abs_pbias(true=self.true, predicted=self.predicted, treat_arrays=False)
 
     def acc(self) -> float:
-        """Anomaly correction coefficient. See Langland_ et al., 2012; Miyakoda_ et al., 1972
+        """Anomaly correction coefficient. See `Langland et al., 2012 <https://doi.org/10.3402/tellusa.v64i0.17531>`_;
+        `Miyakoda_ et al., 1972 <https://doi.org/10.1080/02723646.1972.10642213>`_
         and Murphy_ et al., 1989.
 
         .. math::
             ACC = \\frac{\\sum_{i=1}^{N} \\left( (\\text{predicted}_i - \\overline{\\text{predicted}})(\\text{true}_i - \\overline{\\text{true}}) \\right)}{(N-1) \\cdot \\sigma_{\\text{true}} \\cdot \\sigma_{\\text{predicted}}}
-
-        .. _Langland:
-            https://doi.org/10.3402/tellusa.v64i0.17531
-
-        .. _Miyakoda:
-
-        .. _Murphy:
-            https://doi.org/10.1080/02723646.1972.10642213
 
         Examples
         ---------
@@ -152,16 +145,13 @@ class RegressionMetrics(Metrics):
 
         It detects additive and pro-portional differences in the observed and
         simulated means and vari-ances Moriasi_ et al., 2015. It is overly sensitive
-        to extreme values due to the squared differences_. It can also be used
+        to extreme values due to the squared differences. It can also be used
         as a substitute for R2 to identify the degree to which model predic-tions
         are error-free.
 
         .. math::
             d = 1 - \\frac{\\sum_{i=1}^{N}(e_{i} - s_{i})^2}{\\sum_{i=1}^{N}(\\left | s_{i} - \\bar{e}
              \\right | + \\left | e_{i} - \\bar{e} \\right |)^2}
-
-        .. _differences:
-            Legates and McCabe, 199
 
         .. _Willmott:
             https://doi.org/10.1080/02723646.1981.10642213
@@ -182,16 +172,10 @@ class RegressionMetrics(Metrics):
 
     def aic(self, p=1) -> float:
         """
-        Akaike_ Information Criterion. Modifying from this source_
+        `Akaike_ Information Criterion <https://doi.org/10.1007/978-1-4612-1694-0_15>`_. Modifying from this `source <https://github.com/UBC-MDS/RegscorePy/blob/master/RegscorePy/aic.py>`_
 
         .. math::
             AIC = n \\cdot \\ln\\left(\\frac{\\sum_{i=1}^{n} (\\text{true}_i - \\text{predicted}_i)^2}{n}\\right) + 2p
-
-        .. _Akaike:
-            https://doi.org/10.1007/978-1-4612-1694-0_15
-
-        .. _source:
-            https://github.com/UBC-MDS/RegscorePy/blob/master/RegscorePy/aic.py
 
         Examples
         ---------
@@ -205,13 +189,10 @@ class RegressionMetrics(Metrics):
         return aic(true=self.true, predicted=self.predicted, treat_arrays=False, p=p)
 
     def aitchison(self, center='mean') -> float:
-        """Aitchison distance. used in Zhang_ et al., 2020
+        """Aitchison distance. used in `Zhang et al., 2020 <https://doi.org/10.5194/hess-24-2505-2020>`_
 
         .. math::
[…]
