+++ SeqMetrics/_cls.py
@@ -194,9 +194,7 @@ class ClassificationMetrics(Metrics):
 
     def cross_entropy(self, epsilon=1e-12) -> float:
         """
-        References
-        ----------
-        https://doi.org/10.1007/s10479-005-5724-z
+        calculates `cross entropy <https://doi.org/10.1007/s10479-005-5724-z>`_
 
         Examples
         --------
@@ -219,14 +217,11 @@ class ClassificationMetrics(Metrics):
 
     def accuracy(self, normalize: bool = True) -> float:
         """
-        calculates accuracy
+        calculates `accuracy <https://scikit-learn.org/stable/modules/generated/sklearn.metrics.accuracy_score.html>`_
 
         .. math::
             \\text{Accuracy} = \\frac{\\sum_{i=1}^{N} \\mathbb{1}(true_i = predicted_i)}{N}
 
-        References
-        ----------
-        https://scikit-learn.org/stable/modules/generated/sklearn.metrics.accuracy_score.html
         Examples
         --------
         >>> import numpy as np
@@ -241,11 +236,7 @@ class ClassificationMetrics(Metrics):
 
     def confusion_matrix(self, normalize=False):
         """
-        calculates confusion matrix
-
-        References
-        --------
-        https://scikit-learn.org/stable/modules/generated/sklearn.metrics.confusion_matrix.html
+        calculates `confusion matrix <https://scikit-learn.org/stable/modules/generated/sklearn.metrics.confusion_matrix.html>`_
 
         Examples
         --------
@@ -324,7 +315,8 @@ class ClassificationMetrics(Metrics):
 
     def precision(self, average=None):
         """
-        Returns precision score, also called positive predictive value.
+        Returns `precision <https://scikit-learn.org/stable/modules/generated/sklearn.metrics.precision_score.html>`_ score,
+        also called positive predictive value.
         It is number of correct positive predictions divided by the total
         number of positive predictions.
         .. math::
@@ -335,10 +327,6 @@ class ClassificationMetrics(Metrics):
 
         .. math::
             \\text{Precision}_{\\text{weighted}} = \\frac{\\sum_{i=1}^{N} (TP_i + FN_i) \\cdot \\frac{TP_i}{TP_i + FP_i}}{\\sum_{i=1}^{N} (TP_i + FN_i)}
-
-        References
-        --------
-        https://scikit-learn.org/stable/modules/generated/sklearn.metrics.precision_score.html
 
         Examples
         --------
@@ -356,7 +344,8 @@ class ClassificationMetrics(Metrics):
 
     def recall(self, average=None):
         """
-        It is also called sensitivity or true positive rate. It is
+        It is also called `sensitivity <https://scikit-learn.org/stable/modules/generated/sklearn.metrics.recall_score.html>`_
+        or true positive rate. It is
         number of correct positive predictions divided by the total number of positives
 
         .. math::
@@ -370,10 +359,6 @@ class ClassificationMetrics(Metrics):
 
         .. math::
             \\text{Recall}_{\\text{weighted}} = \\sum_{i=1}^{n} w_i \\cdot \\frac{\\text{TP}_i}{\\text{TP}_i + \\text{FN}_i}
-
-        References
-        --------
-        https://scikit-learn.org/stable/modules/generated/sklearn.metrics.recall_score.html
 
         Examples
         --------
@@ -389,16 +374,13 @@ class ClassificationMetrics(Metrics):
 
     def specificity(self, average=None):
         """
-        It is also called true negative rate or selectivity. It is the probability that
[…]
