+++ SeqMetrics/_rgr.py
@@ -358,7 +358,7 @@ class RegressionMetrics(Metrics):
 
     def covariance(self) -> float:
         """
-        Covariance
+        `Covariance <https://scikit-learn.org/stable/api/sklearn.covariance.html>`_
             .. math::
             Covariance = \\frac{1}{N} \\sum_{i=1}^{N}((e_{i} - \\bar{e}) * (s_{i} - \\bar{s}))
 
@@ -582,7 +582,7 @@ class RegressionMetrics(Metrics):
                        low_flow=low_flow)
 
     def gmae(self) -> float:
-        """ Geometric Mean Absolute Error
+        """ `Geometric Mean Absolute Error <https://doi.org/10.1016/j.isprsjprs.2024.04.015>`_
 
         .. math::
             GMAE = \\left( \\prod_{i=1}^{n} \\left| \\text{true}_i - \\text{predicted}_i \\right| \\right)^{\\frac{1}{n}}
@@ -599,7 +599,9 @@ class RegressionMetrics(Metrics):
         return gmae(true=self.true, predicted=self.predicted, treat_arrays=False)
 
     def gmean_diff(self) -> float:
-        """Geometric mean difference. First geometric mean is calculated for each
+        """
+        `Geometric mean difference <https://www.sciencedirect.com/science/article/abs/pii/S0022316624002281>`_.
+         First geometric mean is calculated for each
         of two samples and their difference is calculated.
 
         .. math::
@@ -675,7 +677,7 @@ class RegressionMetrics(Metrics):
                                        treat_arrays=False)
 
     def inrse(self) -> float:
-        """ Integral Normalized Root Squared Error
+        """ `Integral Normalized Root Squared Error <https://doi.org/10.1016/j.engappai.2023.107559>`_
 
         .. math::
             IN\\text{-}RSE = \\sqrt{\\frac{\\sum_{i=1}^{n} (\\text{true}_i - \\text{predicted}_i)^2}{\\sum_{i=1}^{n} (\\text{true}_i - \\overline{\\text{true}})^2}}
@@ -694,7 +696,8 @@ class RegressionMetrics(Metrics):
         return inrse(true=self.true, predicted=self.predicted, treat_arrays=False)
 
     def irmse(self) -> float:
-        """Inertial RMSE. RMSE divided by standard deviation of the gradient of true.
+        """ `Inertial RMSE <https://link.springer.com/article/10.1007/s11069-008-9299-2>`_.
+        RMSE divided by standard deviation of the gradient of true.
 
         .. math::
             \\text{IRMSE} = \\frac{\\sqrt{\\frac{1}{n} \\sum_{i=1}^{n} \\left( \\text{true}_i - \\text{predicted}_i \\right)^2}}{\\sqrt{\\frac{1}{n-2} \\sum_{i=1}^{n-1} \\left( (\\text{true}_{i+1} - \\text{true}_i) - \\overline{(\\text{true}_{i+1} - \\text{true}_i)} \\right)^2}}
@@ -1320,7 +1323,7 @@ class RegressionMetrics(Metrics):
         return median_abs_error(true=self.true, predicted=self.predicted, treat_arrays=False)
 
     def med_seq_error(self) -> float:
-        """Median Squared Error
+        """`Median Squared Error <https://www.sciencedirect.com/science/article/pii/S2468227620301757>`_
         Same as mse, but it takes median which reduces the impact of outliers.
 
         .. math::
@@ -1411,7 +1414,7 @@ class RegressionMetrics(Metrics):
 
     def msle(self, weights=None) -> float:
[…]
