+++ SeqMetrics/_rgr.py
@@ -475,7 +475,8 @@ class RegressionMetrics(Metrics):
         return decomposed_mse(true=self.true, predicted=self.predicted, treat_arrays=False)
 
     def euclid_distance(self) -> float:
-        """Euclidian distance
+        """ `Euclidian distance <https://scikit-learn.org/stable/modules/generated/sklearn.metrics.pairwise.euclidean_distances.html>`_
+        taken from `this book <https://doi.org/10.1016/B978-0-12-088735-4.50006-7`_.
 
         .. math::
             D = \\sqrt{\\sum_{i=1}^{n} (\\text{true}_i - \\text{predicted}_i)^2}
@@ -890,7 +891,8 @@ class RegressionMetrics(Metrics):
         return kgenp_bound(true=self.true, predicted=self.predicted, treat_arrays=False)
 
     def kl_sym(self) -> Union[float, None]:
-        """Symmetric kullback-leibler divergence
+        """
+        `Symmetric kullback-leibler divergence <https://doi.org/10.1016/j.procs.2018.10.144>`_
 
         .. math::
             \\text{KL}_{\\text{sym}}(P || Q) = \\frac{1}{2} \\sum_{i=1}^{n} \\left( P_i - Q_i \\right) \\left( \\log_2 \\frac{P_i}{Q_i} \\right)
@@ -907,7 +909,8 @@ class RegressionMetrics(Metrics):
         return kl_sym(true=self.true, predicted=self.predicted, treat_arrays=False)
 
     def lm_index(self, obs_bar_p=None) -> float:
-        """Legate-McCabe Efficiency Index.
+        """
+        `Legate-McCabe Efficiency Index <https://doi.org/10.1016/j.cmpb.2023.107737>`_.
         Less sensitive to outliers in the data. The larger, the better
 
         .. math::
@@ -933,7 +936,7 @@ class RegressionMetrics(Metrics):
 
     def maape(self) -> float:
         """
-        Mean Arctangent Absolute Percentage Error
+        `Mean Arctangent Absolute Percentage Error <https://doi.org/10.1016/j.ijforecast.2015.12.003>`_
         Note: result is NOT multiplied by 100
 
         .. math::
@@ -1005,7 +1008,7 @@ class RegressionMetrics(Metrics):
         return mape(true=self.true, predicted=self.predicted, treat_arrays=False)
 
     def mbrae(self, benchmark: np.ndarray = None) -> float:
-        """ Mean Bounded Relative Absolute Error
+        """ `Mean Bounded Relative Absolute Error <https://doi.org/10.1371/journal.pone.0174202>`_
 
         .. math::
             MBRAE = \\frac{1}{n} \\sum_{i=1}^{n} \\frac{| \\text{true}_i - \\text{predicted}_i |}{| \\text{true}_i - \\text{benchmark}_i |}
@@ -1022,8 +1025,8 @@ class RegressionMetrics(Metrics):
         return mbrae(true=self.true, predicted=self.predicted, benchmark=benchmark,
                      treat_arrays=False)
 
-    def mapd(self) -> float:
-        """Mean absolute percentage deviation.
+    def mapd(self) -> float: #ToDo equation not multiplied by 100
+        """ `Mean absolute percentage deviation <https://doi.org/10.1016/j.rinma.2022.100347>`_
 
         .. math::
             MAPD = \\frac{\\sum_{i=1}^{n} \\left| predicted_i - true_i \\right|}{\\sum_{i=1}^{n} \\left| true_i \\right|}
@@ -1085,7 +1088,8 @@ class RegressionMetrics(Metrics):
 
     def max_error(self) -> float:
         """
[…]
