### 2024-07-19 747c0ca
updated gmrae and kgenp_bound references
