### 2024-07-24 f6c8ef7
use .item() on returned value of certain metrics to avoid warning
