+++ requirements.txt
@@ -1,1 +1,1 @@
-numpy<=2.0.1, >=1.17
+numpy<=1.26.4, >=1.17
