+++ .github/workflows/np264.yml
@@ -1,4 +1,5 @@
 name: np264
+# checks on latest numpy version
 on:
   push:
     branches: [ master ]
@@ -25,7 +26,7 @@ jobs:
         python -m pip install --upgrade pip
         python -m pip install flake8 pytest
         pip install -r dev_requirements.txt
-        pip install numpy==2.0.1
+        pip install numpy==1.26.4
     - name: Test with unittest
       run: |
         pip install pytest
