### 2024-07-25 bc3d324
minor bug in testing
