### 2024-07-26 3c21a20
reformat affiliations
