+++ paper/paper.md
@@ -16,9 +16,9 @@ authors:
     orcid: 0000-0002-0031-745X
     affiliation: 2
 affiliations:
- - name: researcher at Environmental AI Rsearch Group, Islamabad, Pakistan
+ - name: Environmental AI Research Group, National Univsersity of Sciences and Technology, Islamabad, Pakistan
    index: 1
- - name: postdoctoral fellow, KAUST
+ - name: King Abdullah University of Science and Technology, Thuwal, Saudi Arabia
    index: 2
 date: 17 Febuary 2024
 bibliography: paper.bib
