+++ paper/paper.bib
@@ -74,7 +74,7 @@ year = {2019}
 
 
  @misc{hydroeval2021,
-  title={hydroeval: an evaluator for streamflow time series in Python},
+  title={hydroeval: an evaluator for streamflow time series in {Python}},
   author={Hallouin, Thibault},
   year={2021},
   publisher={GitHub},
@@ -83,7 +83,7 @@ year = {2019}
 
 
  @article{wade2018hydroerr,
-  title={Hydrostats: A Python Package for Characterizing Errors between Observed and Predicted Time Series},
+  title={Hydrostats: A {Python} Package for Characterizing Errors between Observed and Predicted Time Series},
   author={Roberts, Wade and Williams, Gustavious and Jackson, Elise and Nelson, James and Ames, Daniel},
   year={2021},
   publisher={MDPI},
@@ -96,7 +96,7 @@ year = {2019}
 
 @inproceedings{Wang2020,
 author = {Wang, Ying and Wen, Ming and Liu, Yepang and Wang, Yibo and Li, Zhenming and Wang, Chao and Yu, Hai and Cheung, Shing-Chi and Xu, Chang and Zhu, Zhiliang},
-title = {Watchman: monitoring dependency conflicts for Python library ecosystem},
+title = {Watchman: monitoring dependency conflicts for {Python} library ecosystem},
 year = {2020},
 isbn = {9781450371216},
 publisher = {Association for Computing Machinery},
@@ -114,7 +114,7 @@ series = {ICSE '20}
 
 @inproceedings{Mukherjee2021,
 author = {Mukherjee, Suchita and Almanza, Abigail and Rubio-Gonz\'{a}lez, Cindy},
-title = {Fixing dependency errors for Python build reproducibility},
+title = {Fixing dependency errors for {Python} build reproducibility},
 year = {2021},
 isbn = {9781450384599},
 publisher = {Association for Computing Machinery},
