### 2024-07-26 cb58218
Fix TorchMetrics reference
