+++ paper/paper.bib
@@ -37,14 +37,14 @@ year = {2019}
 }
 
 @article{detlefsen2022torchmetrics,
-  title={Torchmetrics-measuring reproducibility in pytorch},
+  title={TorchMetrics - Measuring Reproducibility in PyTorch},
   author={Detlefsen, Nicki Skafte and Borovec, Jiri and Schock, Justus and Jha, Ananya Harsh and Koker, Teddy and Di Liello, Luca and Stancl, Daniel and Quan, Changsheng and Grechkin, Maxim and Falcon, William},
   journal={Journal of Open Source Software},
   volume={7},
   number={70},
   pages={4101},
   year={2022},
-  doi={10.21105/joss.041012}
+  doi={10.21105/joss.04101}
 }
 
 
