### 2024-07-26 c43c01d
Various minor spelling/syntax/punctuation fixes
