+++ paper/paper.md
@@ -32,11 +32,11 @@ classification models with a comprehensive suite of performance metrics suitable
 for tabular and time series data. Designed for versatility, the library offers 
 112 regression and 22 classification metrics through both functional and 
 class-based APIs. The design of the library ensures seamless integration into 
-various coding environments. The web based GUI of SeqMetrics enhances user 
+various coding environments. The web-based graphical user interface (GUI) of SeqMetrics enhances user 
 accessibility, allowing efficient input through data arrays or file imports. 
 It serves to users of varied programming expertise, offering a user-friendly 
 interface for rigorous model assessment. The library prioritizes computational 
-efficiency and has minimal dependency with straightforward pip installation 
+efficiency and has minimal dependencies with straightforward pip installation 
 from PyPI. Rigorous testing of Seqmetrics ensures robustness, supported by extensive 
 documentation for effective utilization by people from diverse backgrounds. 
 Overall, SeqMetrics bridges the gap in Python's scientific analysis toolkit, 
@@ -53,22 +53,22 @@ Existing Python libraries for performance evaluation offer only a limited number
 For instance, the metrics sub-module from Keras [@chollet2015keras] contains only 24 metrics, 
 while scikit-learn's [@pedregosa2011scikit] metrics module covers 45 metrics. Torchmetrics library, [@detlefsen2022torchmetrics] 
 offers a comprehensive list of over 100 metrics from diverse fields. However, it provides only 48 which are intended
-for 1-dimensional numerical data. Morevoer, the libarary's module based API requires understand of Object Oriented Programming
-and a more simpler functional API does not cover all metrics.
+for 1-dimensional numerical data. Morevoer, the library's module-based API requires understanding of Object Oriented Programming
+and its simpler functional API does not cover all metrics.
 There are also some subject-specific
 libraries. These include NeuralHydrology [@Kratzert2022], hydroeval [@hydroeval2021] and HydroErr [@wade2018hydroerr],
 which address the needs of hydrology and water sciences. However, these resources
-lack more general set of metrics which the experts in these domains often require. 
+lack more general sets of metrics, which the experts in these domains often require. 
 Another limitation of these otherwise valuable libraries is their wide range of dependencies. 
-For example torchmetrics depends upon Pytorch [@paszke2019pytorch] which itself upon several other libraries. This 
-can lead to dependency conflicts due to version constraints complicating the maintenance 
+For example, torchmetrics depends on Pytorch [@paszke2019pytorch], which itself depends on several other libraries. This 
+can lead to dependency conflicts due to version constraints, complicating the maintenance 
 of multiple libraries within a single Python environment [@Wang2020].
 Addressing this gap is imperative to provide 
 researchers a unified platform for 
 comprehensive model evaluation and with minimal requirements, streamlining their computational workflows and 
 enhancing accuracy and reproducibility across various domains [@Mukherjee2021]. The SeqMetrics application addresses 
 the critical need for a robust and versatile toolkit for assessing and comparing 
-regression and classification models performance across a spectrum of domains. 
+regression and classification model performance across a spectrum of domains. 
 With a comprehensive suite of performance metrics for sequential (tabular and 
 time series) data, spanning traditional statistical measures to specialized 
 atmospheric sciences metrics, the software serves as a valuable resource for 
@@ -87,12 +87,12 @@ provided for use with continuous variables. On the other hand, for categorical
 data, there are classification metrics including accuracy, precision, recall, F1 
 score, and the area under the ROC curve, among others. 
 From a programming perspective, the SeqMetrics library employs a modular architecture, 
-offering a functional and class-based APIs for smooth integration across diverse 
[…]
