### 2024-07-26 79fef39
Fix figure captions and references
