### 2024-07-26 3bc0563
Fix installation instructions
