### 2024-07-26 6ac5a8e
correctly identify os in tests
