### 2024-07-26 cfe09c8
convert tensor to flow for test comparison
