### 2024-07-26 bc716b4
minor improvement in docs in tests
