### 2024-07-26 6e1f3db
added log_nse
