+++ app.py
@@ -59,7 +59,7 @@ NAME_REG = {'calculate_all' : 'Calculate All Regression Metrics',
             'kl_sym': 'Symmetric Kullback-leibler Divergence',
             'lm_index': 'Legate-McCabe Efficiency Index',
             'log_prob': 'Logarithmic Probability Distribution',
-            # 'log_nse': 'Log Nash-Sutcliffe model Efficiency',
+            'log_nse': 'Log Nash-Sutcliffe model Efficiency',
             'maape': 'Mean Arctangent Absolute Percentage Error',
             'mbrae': 'Mean Bounded Relative Absolute Error',
             'max_error': 'Maximum Absolute Error',
