+++ paper/paper.md
@@ -9,6 +9,7 @@ tags:
 authors:
   - name: Fazila Rubab
     affiliation: 1
+    orcid: 0009-0004-9040-3459
   - name: Sara Iftikhar
     orcid: 0000-0001-7446-6805
     affiliation: 1
