### 2024-07-26 0516499
added normalized nse
