+++ app.py
@@ -133,6 +133,7 @@ NAME_REG = {'calculate_all' : 'Calculate All Regression Metrics',
             "mape_for_peaks": "MAPE for peaks",
             "legates_coeff_eff": "Legates Coefficient of Efficiency",
             "manhattan_distance": "Manhattan distance",
+            'norm_nse': "Normalized Nash-Sutcliffe Efficiency",
             }
 
 NAME_CLS = {'accuracy' : 'Accuracy',
