+++ tests/test_rgr.py
@@ -1839,8 +1839,9 @@ class Test_Torch_metrics(unittest.TestCase):
             new_critical_success_index = metrics.critical_success_index()
             csi = CriticalSuccessIndex(0.5)
             torch_csi = tensor_to_float(csi(torch.tensor(p11), torch.tensor(t11)))
-            self.assertAlmostEqual(new_critical_success_index, torch_csi)
+            self.assert_almost_equal(new_critical_success_index, torch_csi, "critical success index")
         return
+
     def test_critical_success_index_func(self):
         try:
             import torch
@@ -1852,7 +1853,7 @@ class Test_Torch_metrics(unittest.TestCase):
             new_critical_success_index = sm_critical_success_index(t11, p11)
             csi = CriticalSuccessIndex(0.5)
             torch_csi = tensor_to_float(csi(torch.tensor(p11), torch.tensor(t11)))
-            self.assertAlmostEqual(new_critical_success_index, torch_csi)
+            self.assert_almost_equal(new_critical_success_index, torch_csi, "critical success index")
         return
 
     def test_kl_divergence_cls(self):
@@ -1866,7 +1867,7 @@ class Test_Torch_metrics(unittest.TestCase):
             new_kl_divergence = metrics.kl_divergence()
             kl_div = KLDivergence()
             torch_kl_div = tensor_to_float(kl_div(torch.tensor(p11).reshape(1,-1), torch.tensor(t11).reshape(1,-1)))
-            self.assertAlmostEqual(new_kl_divergence, torch_kl_div)
+            self.assert_almost_equal(new_kl_divergence, torch_kl_div, "kullback leibler divergence")
         return
 
     def test_kl_divergence_func(self):
@@ -1880,7 +1881,7 @@ class Test_Torch_metrics(unittest.TestCase):
             new_kl_divergence = sm_kl_divergence(t11, p11)
             kl_div = KLDivergence()
             torch_kl_div = tensor_to_float(kl_div(torch.tensor(p11).reshape(1,-1), torch.tensor(t11).reshape(1,-1)))
-            self.assertAlmostEqual(new_kl_divergence, torch_kl_div)
+            self.assert_almost_equal(new_kl_divergence, torch_kl_div, "kullback leibler divergence")
         return
 
     def test_log_cosh_error_cls(self):
@@ -1894,7 +1895,7 @@ class Test_Torch_metrics(unittest.TestCase):
             new_log_cosh_error = metrics.log_cosh_error()
             lg_cosh_err = LogCoshError()
             torch_lg_cosh_err = tensor_to_float(lg_cosh_err(torch.tensor(p11), torch.tensor(t11)))
-            self.assertAlmostEqual(new_log_cosh_error, torch_lg_cosh_err)
+            self.assert_almost_equal(new_log_cosh_error, torch_lg_cosh_err, "log cosh error")
         return
 
     def test_log_cosh_error_func(self):
@@ -1909,7 +1910,7 @@ class Test_Torch_metrics(unittest.TestCase):
             lg_cosh_err = LogCoshError()
             torch_lg_cosh_err = tensor_to_float(lg_cosh_err(torch.tensor(p11), torch.tensor(t11)))
 
-            self.assertAlmostEqual(new_log_cosh_error, torch_lg_cosh_err)
+            self.assert_almost_equal(new_log_cosh_error, torch_lg_cosh_err, "log cosh error")
         return
 
     def test_minkowski_distance_cls(self):
@@ -1923,7 +1924,7 @@ class Test_Torch_metrics(unittest.TestCase):
             new_minkowski_distance = metrics.minkowski_distance()
             mink_dist = MinkowskiDistance(1)
[…]
