+++ SeqMetrics/_rgr.py
@@ -761,16 +761,23 @@ class RegressionMetrics(Metrics):
     def kge(self):
         """
         Kling-Gupta Efficiency following `Gupta_ et al. 2009 <https://doi.org/10.1016/j.jhydrol.2009.08.003>`_.
+        This error considers `correlation`, `variability` and `mean` difference/error.
 
         .. math::
             \\text{KGE} = 1 - \\sqrt{(r - 1)^2 + (\\alpha - 1)^2 + (\\beta - 1)^2}
         .. math::
-            \\r = \\frac{\\sum_{i=1}^{N} ( \\text{true}_i - \\bar{\\text{true}} ) ( \\text{predicted}_i - \\bar{\\text{predicted}} )}{\\sqrt{\\sum_{i=1}^{N} ( \\text{true}_i - \\bar{\\text{true}} )^2} \\sqrt{\\sum_{i=1}^{N} ( \\text{predicted}_i - \\bar{\\text{predicted}} )^2}}
-        .. math::
             \\alpha = \\frac{\\sigma_{\\text{predicted}}}{\\sigma_{\\text{true}}}
         .. math::
-            \\beta = \\frac{\\mu_{\\text{predicted}}}{\\mu_{\\text{true}}}
+            \\beta = \\frac{\\mu_{\\text{predicted}}}{\\mu_{\\text{true}}}            
+            
+        In this equation, alpha accounts for the variability (standard deviation), beta accounts for 
+        the mean difference and r accounts for the correlation between the true and predicted values.
+        This equation can also be written as belowS            
+            
+        .. math::
+            \\text{KGE} = \\frac{\\sum_{i=1}^{N} ( \\text{true}_i - \\bar{\\text{true}} ) ( \\text{predicted}_i - \\bar{\\text{predicted}} )}{\\sqrt{\\sum_{i=1}^{N} ( \\text{true}_i - \\bar{\\text{true}} )^2} \\sqrt{\\sum_{i=1}^{N} ( \\text{predicted}_i - \\bar{\\text{predicted}} )^2}}
 
+            
         Examples
         ---------
         >>> import numpy as np
@@ -3056,15 +3063,23 @@ def kge(true,
         **treat_arrays_kws):
     """
     Kling-Gupta Efficiency following `Gupta_ et al. 2009 <https://doi.org/10.1016/j.jhydrol.2009.08.003>`_.
+    This error considers `correlation`, `variability` and `mean` difference/error.
 
     .. math::
         \\text{KGE} = 1 - \\sqrt{(r - 1)^2 + (\\alpha - 1)^2 + (\\beta - 1)^2}
-    .. math::
-        \\r = \\frac{\\sum_{i=1}^{N} ( \\text{true}_i - \\bar{\\text{true}} ) ( \\text{predicted}_i - \\bar{\\text{predicted}} )}{\\sqrt{\\sum_{i=1}^{N} ( \\text{true}_i - \\bar{\\text{true}} )^2} \\sqrt{\\sum_{i=1}^{N} ( \\text{predicted}_i - \\bar{\\text{predicted}} )^2}}
+
     .. math::
         \\alpha = \\frac{\\sigma_{\\text{predicted}}}{\\sigma_{\\text{true}}}
     .. math::
-        \\beta = \\frac{\\mu_{\\text{predicted}}}{\\mu_{\\text{true}}}
+        \\beta = \\frac{\\mu_{\\text{predicted}}}{\\mu_{\\text{true}}}        
+        
+    In this equation, alpha accounts for the variability (standard deviation), beta accounts for 
+    the mean difference and r accounts for the correlation between the true and predicted values.
+    This equation can also be written as belowS        
+        
+    .. math::
[…]
