### 2024-07-26 #6450 @AtrCheema
HI @mstimberg Thanks for the PR and suggestions. As per your suggestions, we have completed the following tasks

-> [made suggested edit in paper](https://github.com/AtrCheema/SeqMetrics/commit/e541f07c73375409abbd49b49d3aee3a980a3431)
-> [added ORCID info of first author](https://github.com/AtrCheema/SeqMetrics/commit/6c74a9acd4eecac030d0232e30bbfcdca857bd3f)
-> [released package on pypi](https://pypi.org/project/SeqMetrics/2.0.0/)
-> [archived release on Zenodo](https://zenodo.org/doi/10.5281/zenodo.12958901)

Please let me know if we have to anything else.
