### 2024-07-26 #6450 @mstimberg
Many thanks for the update, @AtrCheema.

Could you please fix the metadata on your Zenodo archive manually (no need for a new release/version), in order to:
- Match authors and titles to the one of the JOSS paper (also include the ORCIDs here)
- Fix the licence to match your code licence

Thanks !
