### 2024-07-26 944e674
added zenodo badge and corrected  author affiliations
