### 2024-07-26 #6450 @mstimberg
> @mstimberg I just noticed a small mistake in the author affiliation. The affiliation of first two authors is [`Environmental AI Research Group, Islamabad, Pakistan`](https://environmentalai.github.io/) and it is not part of National University of Science and Technology. The later is also located in Islamabad, but has nothing to do with affiliation of first two authors. I have corrected it in [paper.md](https://github.com/AtrCheema/SeqMetrics/commit/944e6748fe7eedd92a304a6a3e8829e5336b587f). Should I update the release on Zenodo as well?

Apologies for introducing this mistake… Yes, please edit the Zenodo archive so that it shows the same affiliations as the paper. Also please edit the title so that it no longer contains the name of the GitHub repository,
