### 2024-07-28 #6450 @AtrCheema
@mstimberg

> Also please edit the title so that it no longer contains the name of the GitHub repository

If I understood it correctly, it means the word `SeqMetrics` can not appear as title of the paper?
