### 2024-07-28 #6450 @mstimberg
> If I understood it correctly, it means the word `SeqMetrics` can not appear as title of the paper?

No, my comment was referring to the Zenodo archive which should have the same title as the paper, i.e. please remove the "AtrCheema/SeqMetrics" at the beginning. "
