### 2024-07-28 #6450 @AtrCheema
@mstimberg Thanks for the prompt response. I have [updated the Zenodo archive](https://zenodo.org/records/12964024) with correct title and the latest release has updated paper.md file. However on Zenodo itself, I could not add the affiliation of first two authors explicitly, probably because [Zenodo relies on ROR](https://support.zenodo.org/help/en-gb/1-upload-deposit/22-can-you-add-update-my-affiliation-and-or-funder#:~:text=Zenodo%20relies%20on%20the%20Research,ror.org%2Fregistry%2F.) to search for organizations and the organization of first two authors is not enlisted in the database of ROR yet.
