### 2024-07-29 #6450 @mstimberg
@AtrCheema Thanks for the changes, but as I mentioned earlier, there was no need to do a new release, since only Zenodo metadata and the paper changed, not the code itself. Now we are in the unfortunate situation where the release is `latest` and not a version number. Also, the licence needs manual fixing again. I don't know that the easiest solution for you is here – maybe do a v2.0.1 release?
Regarding the affiliations: this is fine. Maybe add a note as suggested by the Zenodo docs: https://help.zenodo.org/docs/deposit/describe-records/descriptions/#note ?
