### 2024-07-29 #6450 @mstimberg
Yes, you are right – I stumbled a bit over the fact that on Zenodo it is now displayed as not being the latest version, but this shouldn't be an issue. I'll have a final look over the manuscript and then will hand things over to a topic editor :+1:
