### 2024-07-29 16c2ee9
Some additional minor editing fixes
