### 2024-07-29 #6450 @mstimberg
@AtrCheema I had a last final look over the paper and noticed a few minor issues. Please have a look at  AtrCheema/SeqMetrics#5 and merge if you agree. After that I will recommend acceptance and hand over to the topic editor. Thanks for your patience!
