### 2024-07-30 #6450 @danielskatz
@AtrCheema - I'm suggesting small changes in https://github.com/AtrCheema/SeqMetrics/pull/6. Please merge this, or let me know what you disagree with, then we can proceed.  Also, I notice that there is no acknowledgements section, so I want to confirm that you do not have any funding sources or other things that should be mentioned in such a section.
