### 2024-07-30 #6450 @editorialbot
:wave: @openjournals/csism-eics, this paper is ready to be accepted and published.
